### 2024-11-07 #7436 @editorialbot
[PRE REVIEW]: Underworld3: Mathematically Self-Describing Modelling in Python for Desktop, HPC and Cloud
**Submitting author:** @lmoresi (<a href="http://orcid.org/0000-0003-3685-174X">Louis Moresi</a>)
**Repository:** https://github.com/underworldcode/underworld3
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 0.99.0b
