+++ tests/Ex_IndexSwarmVariable.py
@@ -1,0 +1,572 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# In[1]:
+
+
+import underworld3 as uw
+import numpy as np
+import sympy
+
+import pyvista as pv
+import underworld3.visualisation as vis
+
+
+# In[2]:
+
+
+from posixpath import pardir
+import petsc4py.PETSc as PETSc
+
+import os
+import warnings
+from typing import Optional, Tuple
+
+import underworld3 as uw
+from underworld3.utilities._api_tools import Stateful
+from underworld3.utilities._api_tools import uw_object
+
+import underworld3.timing as timing
+from underworld3.swarm import SwarmVariable
+
+class IndexSwarmVariable_B(SwarmVariable):
+    """
+    The IndexSwarmVariable is a class for managing material point
+    behaviour. The material index variable is rendered into a
+    collection of masks each representing the extent of one material
+    """
+
+    @timing.routine_timer_decorator
+    def __init__(
+        self,
+        name,
+        swarm,
+        indices=1,
+        npoints=5,
+        proxy_degree=1,
+        proxy_continuous=True,
+    ):
+        self.indices = indices
+        self.nnn = npoints
+
+        # These are the things we require of the generic swarm variable type
+        super().__init__(
+            name,
+            swarm,
+            size=1,
+            vtype=None,
+            dtype=int,
+            _proxy=False,
+        )
+        """
+        vtype = (None,)
+        dtype = (float,)
+        proxy_degree = (1,)
+        proxy_continuous = (True,)
+        _register = (True,)
+        _proxy = (True,)
+        _nn_proxy = (False,)
+        varsymbol = (None,)
+        rebuild_on_cycle = (True,)
+        """
+        # The indices variable defines how many "level set" maps we create as components in the proxy variable
+
+        import sympy
+
+        self._MaskArray = sympy.Matrix.zeros(1, self.indices)
+        self._meshLevelSetVars = [None] * self.indices
+
+        for i in range(indices):
+            self._meshLevelSetVars[i] = uw.discretisation.MeshVariable(
+                name + R"^{[" + str(i) + R"]}",
+                self.swarm.mesh,
+                num_components=1,
+                degree=proxy_degree,
+                continuous=proxy_continuous,
+            )
+            self._MaskArray[0, i] = self._meshLevelSetVars[i].sym[0, 0]
+
+        return
+
+    # This is the sympy vector interface - it's meaningless if these are not spatial arrays
+    @property
+    def sym(self):
+        return self._MaskArray
+
+    @property
+    def sym_1d(self):
+        return self._MaskArray
+
+    # We can  also add a __getitem__ call to access each mask
+
+    def __getitem__(self, index):
+        return self.sym[index]
+
+    def createMask(self, funcsList):
+        """
+        This creates a masked sympy function of swarm variables required for Underworld's solvers
+        """
+
+        if not isinstance(funcsList, (tuple, list)):
+            raise RuntimeError("Error input for createMask() - wrong type of input")
+
+        if len(funcsList) != self.indices:
+            raise RuntimeError("Error input for createMask() - wrong length of input")
+
+        symo = sympy.simplify(0)
+        for i in range(self.indices):
+            symo += funcsList[i] * self._MaskArray[i]
[…]
