### 2024-11-20 202a89a
Handles changes in swarm.dm.getField interface in PETSc >= 3.22.0
