+++ src/underworld3/swarm.py
@@ -960,14 +960,14 @@ class Swarm(Stateful, uw_object):
         self.dm.finalizeFieldRegister()
         self.dm.addNPoints(newp_coords.shape[0] + 1)
 
-        cellid = self.dm.getField("DMSwarm_cellid")
-        coords = self.dm.getField("DMSwarmPIC_coor").reshape((-1, self.dim))
+        cellid = self._dm_get_field("DMSwarm_cellid")
+        coords = self._dm_get_field("DMSwarmPIC_coor").reshape((-1, self.dim))
 
         coords[...] = newp_coords[...]
         cellid[:] = newp_cells[:]
 
-        self.dm.restoreField("DMSwarmPIC_coor")
-        self.dm.restoreField("DMSwarm_cellid")
+        self._dm_restore_field("DMSwarmPIC_coor")
+        self._dm_restore_field("DMSwarm_cellid")
 
         ## Now make a series of copies to allow the swarm cycling to
         ## work correctly (if required)
@@ -998,8 +998,8 @@ class Swarm(Stateful, uw_object):
 
             self.dm.addNPoints(swarm_new_size - swarm_orig_size)
 
-            cellid = self.dm.getField("DMSwarm_cellid")
-            coords = self.dm.getField("DMSwarmPIC_coor").reshape((-1, self.dim))
+            cellid = self._dm_get_field("DMSwarm_cellid")
+            coords = self._dm_get_field("DMSwarmPIC_coor").reshape((-1, self.dim))
 
             coords[...] = (
                 all_local_coords[...]
@@ -1010,8 +1010,8 @@ class Swarm(Stateful, uw_object):
             )
             cellid[:] = all_local_cells[:, 0]
 
-            self.dm.restoreField("DMSwarmPIC_coor")
-            self.dm.restoreField("DMSwarm_cellid")
+            self._dm_restore_field("DMSwarmPIC_coor")
+            self._dm_restore_field("DMSwarm_cellid")
 
             ## Now set the cycle values
 
@@ -1084,14 +1084,14 @@ class Swarm(Stateful, uw_object):
         self.dm.finalizeFieldRegister()
         self.dm.addNPoints(npoints=npoints)
 
-        cellid = self.dm.getField("DMSwarm_cellid")
-        coords = self.dm.getField("DMSwarmPIC_coor").reshape((-1, self.dim))
+        cellid = self._dm_get_field("DMSwarm_cellid")
+        coords = self._dm_get_field("DMSwarmPIC_coor").reshape((-1, self.dim))
 
         coords[swarm_size::, :] = valid_coordinates[:, :]
         cellid[swarm_size::] = valid_cells[:]
 
-        self.dm.restoreField("DMSwarmPIC_coor")
-        self.dm.restoreField("DMSwarm_cellid")
+        self._dm_restore_field("DMSwarmPIC_coor")
+        self._dm_restore_field("DMSwarm_cellid")
 
         # Here we update the swarm cycle values as required
 
@@ -1466,7 +1466,7 @@ class Swarm(Stateful, uw_object):
             # add to de-access list to rewind this later
             deaccess_list.append(var)
             # grab numpy object, setting read only if necessary
-            var._data = self.dm.getField(var.clean_name).reshape(
+            var._data = self._dm_get_field(var.clean_name).reshape(
                 (-1, var.num_components)
             )
             assert var._data is not None
@@ -1509,7 +1509,7 @@ class Swarm(Stateful, uw_object):
                     if var not in writeable_vars:
                         var._data.flags.writeable = var._old_data_flag
                     var._data = None
-                    self.em_swarm.dm.restoreField(var.clean_name)
+                    self.em_swarm._dm_restore_field(var.clean_name)
[…]
