### 2024-11-27 2d3a089
Mostly cosmetic changes for the solver feedback that we advertise in the JOSS paper
Expressions should expand in the "where:" section of the solver self-description and a couple of other spots.

Added a tensor contraction function as well, but that is probably not of interest to anyone.
