+++ src/underworld3/cython/petsc_generic_snes_solvers.pyx
@@ -132,7 +132,9 @@ class SolverBaseClass(uw_object):
         from IPython.display import Latex, Markdown, display
         from textwrap import dedent
 
-        display(Markdown(fr"This solver is formulated in {self.mesh.dim} dimensions"))
+        display(Markdown(fr"### Boundary Conditions"))
+
+        display(Markdown(fr"This solver is formulated as {self.mesh.dim} dimensional problem with a {self.mesh.cdim} dimensional mesh"))
 
         return
 
@@ -544,10 +546,11 @@ class SNES_Scalar(SolverBaseClass):
 
         # Should del these when finished
 
-
         self.petsc_fe_u = PETSc.FE().createDefault(mesh.dim, 1, mesh.isSimplex, mesh.qdegree, "private_{}_".format(self.petsc_options_prefix), PETSc.COMM_SELF,)
         self.petsc_fe_u_id = self.dm.getNumFields()
         self.dm.setField( self.petsc_fe_u_id, self.petsc_fe_u )
+        self.petsc_fe_u.setName("_scalar_unknown_")
+
         self.is_setup = False
 
         if self.verbose:
@@ -655,6 +658,7 @@ class SNES_Scalar(SolverBaseClass):
         else:
             if verbose and uw.mpi.rank == 0:
                 print(f"SNES_Scalar ({self.name}): Pointwise functions need to be built", flush=True)
+
 
 
         mesh = self.mesh
@@ -928,7 +932,7 @@ class SNES_Scalar(SolverBaseClass):
 
         # feedback on this instance
         display(
-            Markdown(f"**Poisson system solver**"),
+            Markdown(f"### Poisson system solver"),
             Markdown(f"Primary problem: "),
             Latex(eqF1), Latex(eqf0),
         )
@@ -944,7 +948,20 @@ class SNES_Scalar(SolverBaseClass):
                 expr._object_viewer(description=False)
 
 
-        display(Markdown(fr"This solver is formulated in {self.mesh.dim} dimensions"))
+        display(
+            Markdown(fr"#### Boundary Conditions"),)
+
+        bc_table = "| Type   | Boundary | Expression | \n"
+        bc_table += "|:------------------------ | -------- | ---------- | \n"
+
+        for bc in self.essential_bcs:
+            bc_table += f"| **{bc.type}** | {bc.boundary} | ${sympy.latex(bc.fn.T)}  $ | \n"
+        for bc in self.natural_bcs:
+                bc_table += f"| **{bc.type}** | {bc.boundary} | ${sympy.latex(bc.fn_f.T)}  $ | \n"
+
+        display(Markdown(bc_table))
+
+        display(Markdown(fr"This solver is formulated as a {self.mesh.dim} dimensional problem with a {self.mesh.cdim} dimensional mesh"))
 
 
 
@@ -1121,6 +1138,7 @@ class SNES_Vector(SolverBaseClass):
         self.petsc_fe_u = PETSc.FE().createDefault(mesh.dim, mesh.dim, mesh.isSimplex, mesh.qdegree, "private_{}_u_".format(self.petsc_options_prefix), PETSc.COMM_SELF)
         self.petsc_fe_u_id = self.dm.getNumFields()
         self.dm.setField( self.petsc_fe_u_id, self.petsc_fe_u )
+        self.petsc_fe_u.setName("_vector_unknown_")
 
         self.dm.createDS()
 
@@ -1598,7 +1616,7 @@ class SNES_Vector(SolverBaseClass):
 
         # feedback on this instance
         display(
-            Markdown(f"**Vector poisson solver**"),
+            Markdown(f"### Vector poisson solver"),
             Markdown(f"Primary problem: "),
             Latex(eqF1), Latex(eqf0),
[…]
