### 2024-12-06 92a94f5
Restore .environment.yml
