### 2024-12-17 e26b413
Fixes for #274 - conda environment.yml bugs.
- environment.yml need numpy<2
- setup.cfg calls auto version label, half loading uw3 - remove for now.

MUST pip install with '--no-build-isolation' argument for now.
