+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = underworld3
-version = attr: underworld3.__version__
+#version = attr: underworld3.__version__
 
 #[build_ext]
 #inplace = True
