+++ src/underworld3/function/__init__.py
@@ -15,6 +15,7 @@ from .expressions import unwrap as fn_unwrap
 from .expressions import substitute_expr as fn_substitute_one_expression
 from .expressions import is_constant_expr as fn_is_constant_expr
 from .expressions import extract_expressions as fn_extract_expressions
+from .expressions import extract_expressions as fn_extract_expressions_and_functions
 
 # from .expressions import UWconstant_expression as constant
 
@@ -44,9 +45,9 @@ def derivative(expression, variable, evaluate=True):
                 for t in x.as_ordered_terms():
                     d += sympy.diff(t, variable, evaluate=False)
 
-                return d
+                return
 
-            derivative = expression.applyfunc(f)
+            derivative = expression.applyfunc(f).as_mutable()
 
             # f = lambda x: sympy.diff(
             #         x, variable, evaluate=False
