### 2024-12-18 d1c7542
F0 as expression - copy instead of rebuild
We need to be careful to retain the original expression objects in solver F0 and F1 otherwise we lose the lazy evaluation property. This is for Stokes bodyforce and Stokes F0. Others have not been checked.
