### 2024-12-18 bc47ecd
RegionalSpherical redefinition of class
This is now a SW-NE specified box on the sphere. Clearly this needs some more work to be complete, particularly in coordinate transformations
