+++ src/underworld3/meshing.py
@@ -2750,7 +2750,11 @@ def CubedSphere(
 def RegionalSphericalBox(
     radiusOuter: float = 1.0,
     radiusInner: float = 0.547,
-    numElements: int = 5,
+    SWcorner=[-45,-45],
+    NEcorner=[+45,+45],
+    numElementsLon: int = 5,
+    numElementsLat: int = 5,
+    numElementsDepth: int = 5,
     degree: int = 1,
     qdegree: int = 2,
     simplex: bool = False,
@@ -2769,14 +2773,66 @@ def RegionalSphericalBox(
         East = 5
         West = 6
 
+    lt_min = np.radians(SWcorner[1])
+    lt_max = np.radians(NEcorner[1])
+    ln_min = np.radians(SWcorner[0])
+    ln_max = np.radians(NEcorner[0])
 
-    r1 = radiusInner / np.sqrt(3)
-    r2 = radiusOuter / np.sqrt(3)
+    p2 = (  radiusOuter * np.cos(lt_max) * np.cos(ln_max),
+            radiusOuter * np.cos(lt_max) * np.sin(ln_max),
+            radiusOuter * np.sin(lt_max)
+    )
+
+    p3 = (  radiusOuter * np.cos(lt_max) * np.cos(ln_min),
+            radiusOuter * np.cos(lt_max) * np.sin(ln_min),
+            radiusOuter * np.sin(lt_max)
+        )
+
+    p4 = (  radiusOuter * np.cos(lt_min) * np.cos(ln_min),
+            radiusOuter * np.cos(lt_min) * np.sin(ln_min),
+            radiusOuter * np.sin(lt_min)
+    )
+
+    p5 = (  radiusOuter * np.cos(lt_min) * np.cos(ln_max),
+            radiusOuter * np.cos(lt_min) * np.sin(ln_max),
+            radiusOuter * np.sin(lt_min)
+    )
+
+    p6 = (  radiusInner * np.cos(lt_max) * np.cos(ln_max),
+            radiusInner * np.cos(lt_max) * np.sin(ln_max),
+            radiusInner * np.sin(lt_max)
+    )
+
+    p7 = (  radiusInner * np.cos(lt_max) * np.cos(ln_min),
+            radiusInner * np.cos(lt_max) * np.sin(ln_min),
+            radiusInner * np.sin(lt_max)
+        )
+
+    p8 = (  radiusInner * np.cos(lt_min) * np.cos(ln_min),
+            radiusInner * np.cos(lt_min) * np.sin(ln_min),
+            radiusInner * np.sin(lt_min)
+    )
+
+    p9 = (  radiusInner * np.cos(lt_min) * np.cos(ln_max),
+            radiusInner * np.cos(lt_min) * np.sin(ln_max),
+            radiusInner * np.sin(lt_min)
+    )
+
+
+    # lat_south = np.radians(centralLatitude - latitudeExtent/2)
+    # lat_north = np.radians(centralLatitude + latitudeExtent/2)
+
+    # ss = min(longitudeExtent / 90, 1.99) * np.cos(lat_south)/np.cos(np.pi/4)
+    # sn = min(longitudeExtent / 90, 1.99) * np.cos(lat_north)/np.cos(np.pi/4)
+    # t = min(latitudeExtent / 90, 1.99)
+
+    # r1 = radiusInner / np.sqrt(3)
+    # r2 = radiusOuter / np.sqrt(3)
 
     if filename is None:
         if uw.mpi.rank == 0:
             os.makedirs(".meshes", exist_ok=True)
[…]
