### 2025-01-06 8bbecc5
Create environment-macos.yml
It just seems inevitable that we have to have different builds for different machines at this point
