+++ environment-macos.yml
@@ -1,0 +1,31 @@
+name: uw3-macos
+channels:
+  - conda-forge
+dependencies:
+  - python=3.12
+  - compilers
+  - mpich
+  - petsc
+  - petsc4py
+  - cython=3.*
+  - mpmath<=1.3
+  - mesalib
+  - numpy
+  - scipy
+  - mpi4py
+  - h5py
+  - sympy
+  - pytest
+  - ipython
+  - pyvista
+  - pip
+  - pip:
+      - gmsh
+      - gmsh-api
+      - jupyterlab
+      - jupytext
+      - pygments
+      - ipywidgets
+      - trame
+      - trame-vtk
+      - trame-vuetify
