### 2025-01-16 3186550
Gather the CC used by PETSc for UW3 extensions.
Hopefully this is a fix for effiectively getting mpicc.
Need to test using conda next.
