### 2025-01-16 8c70ecf
RBF interpolation / kdtree add options
Adding function to collect data to kdtree points or from kdtree points in the RBF interpolator functionality. We tend to use both depending on the way we want to map between point clouds and so both ways need a general interface.
