+++ src/underworld3/kdtree.pyx
@@ -1,3 +1,4 @@
+from types import WrapperDescriptorType
 import underworld3
 import underworld3 as uw
 import underworld3.timing as timing
@@ -190,6 +191,9 @@ cdef class KDTree:
 ## A general point-to-point rbf interpolator here
 ## NOTE this is not using cython optimisation for numpy
 
+    # For backward compatibility, default for the rbf_interpolator function
+    # is the _from_kdtree version
+
     def rbf_interpolator_local(self,
             coords,
             data,
@@ -197,11 +201,24 @@ cdef class KDTree:
             verbose = False,
         ):
 
+        return self.rbf_interpolator_local_from_kdtree(
+            coords, data, nnn, verbose,
+        )
+
+
+    def rbf_interpolator_local_from_kdtree(self,
+            coords,
+            data,
+            nnn = 4,
+            verbose = False,
+        ):
+
         '''
-        An inverse (squared) distance weighted mapping of a numpy array from one
-        set of coordinates to another. This assumes all points are local to the
-        same processor. If that is not the case, it is best to use a particle swarm
-        to migrate data.
+        An inverse (squared) distance weighted mapping of a numpy array from the
+        set of coordinates defined by the kd-tree to the set of input points specified.
+        This assumes all points are local to the same processor.
+        If that is not the case, it is best to use a particle swarm
+        to manage the distributed data.
         '''
 
         if coords.shape[1] != self.points.shape[1]:
@@ -216,8 +233,11 @@ cdef class KDTree:
         closest_n, distance_n = self.find_closest_n_points(nnn, coords_contiguous)
 
         num_local_points = coords.shape[0]
-        data_size = data.shape[1]
-
+        try:
+            data_size = data.shape[1]
+        except IndexError:
+            data_size = 1
+            data = data.reshape(-1,1)
         Values = np.zeros((num_local_points, data_size))
         Weights = np.zeros((num_local_points, 1))
 
@@ -244,5 +264,74 @@ cdef class KDTree:
         del closest_n
         del distance_n
         del Weights
+
+        return Values
+
+    def rbf_interpolator_local_to_kdtree(self,
+                    coords,
+                    data,
+                    nnn = 4,
+                    verbose = False,
+                    weights = None
+                ):
+
+        '''
+        An inverse (squared) distance weighted mapping of a numpy array to the
+        set of coordinates defined by the kd-tree from the set of input points specified.
+        This assumes all points are local to the same processor.
+        If that is not the case, it is sensible to use a particle swarm
+        to manage the distributed data.
+        '''
+
+        if coords.shape[1] != self.points.shape[1]:
+            raise RuntimeError(f"Interpolation coordinates dimensionality ({coords.shape[1]}) is different to kD-tree dimensionality ({self.points.shape[1]}).")
+        nInput = coords.shape[0]
+
+        if data.shape[0] != coords.shape[0]:
+                raise RuntimeError(f"Data does not match coords size array ({data.shape[0]}) v ({coords.shape[0]}).")
+
+        coords_contiguous = np.ascontiguousarray(coords)
+
+        closest_n, distance_n = self.find_closest_n_points(nnn, coords_contiguous)
+
+        num_local_points = self.points.shape[0]
+        try:
+            data_size = data.shape[1]
+        except IndexError:
+            data_size = 1
+            data = data.reshape(-1,1)
+
+        Values = np.zeros((num_local_points, data_size))
+        Weights = np.zeros((num_local_points, 1))
+
+        if verbose and uw.mpi.rank == 0:
+            print(f"Mapping values  ... start", flush=True)
+
+        epsilon = 1.0e-9
[…]
