### 2025-01-16 cdd5fb0
set CC env variable (#266)
* set CC env variable

Fixes issues with finding mpi

* Gather the CC used by PETSc for UW3 extensions.

Hopefully this is a fix for effiectively getting mpicc.
Need to test using conda next.

---------

Co-authored-by: Julian Giordani <julesghub@users.noreply.github.com>
