### 2025-01-17 e9471bd
Make dev experience better
1. Add version information back to install, it went missing in recent
changes.

2. Remove git revision in utils.auditor - this can be done better when we
move to pyproject.toml

3. Update setup.cfg to be numpy<2
