+++ setup.cfg
@@ -9,7 +9,7 @@ name = underworld3
 zip_safe = True
 install_requires = 
   xxhash
-  numpy >=1.20
+  numpy <2
   scipy
   cython
   sympy >=1.12
