### 2025-01-22 123d681
fixes for the bodyforce sign change - tests running again
