### 2025-01-22 #7436 @jakelangham
Hi @diehlpk, looks interesting, but I am not really a Python programmer so wouldn't be able to give this a thorough review.
