### 2025-01-22 07b0a72
Attempt to align the 2 conda environments.
More testing to come.
