+++ environment-macos.yml
@@ -2,15 +2,15 @@ name: uw3-macos
 channels:
   - conda-forge
 dependencies:
-  - python=3.12
+  - python=3.11
   - compilers
   - mpich
-  - petsc
-  - petsc4py
+  - petsc=3.21.5
+  - petsc4py=3.21.5
   - cython=3.*
   - mpmath<=1.3
   - mesalib
-  - numpy
+  - numpy<2
   - scipy
   - mpi4py
   - h5py
@@ -29,3 +29,4 @@ dependencies:
       - trame
       - trame-vtk
       - trame-vuetify
+      - pint
