### 2025-01-23 f76bb50
more tweaks to environment.yaml
