+++ environment.yml
@@ -4,10 +4,8 @@ channels:
 dependencies:
   - python=3.11
   - compilers
-  - openmpi
-  - openmpi-mpicc
-  - openmpi-mpicxx
-  - petsc=3.21.5 # Fix for now (errors with hdf5 / uw3)
+  - mpich
+  - petsc=3.21.5
   - petsc4py=3.21.5
   - cython=3.*
   - mpmath<=1.3
@@ -22,7 +20,7 @@ dependencies:
   - sympy
   - ipython
   - mpi4py
-  - h5py
+  - h5py=*=mpi*
   - trame
   - ipywidgets
   - pip
