### 2025-01-23 45a63ba
Unifying conda environments for mac and intel
* Removing `environment-macos.yml`
* Adding Lior's packages
* Comment on memory leak in parallel test (to revisit)
