+++ environment-macos.yml
@@ -1,32 +1,0 @@
-name: uw3-macos
-channels:
-  - conda-forge
-dependencies:
-  - python=3.11
-  - compilers
-  - mpich
-  - petsc=3.21.5
-  - petsc4py=3.21.5
-  - cython=3.*
-  - mpmath<=1.3
-  - mesalib
-  - numpy<2
-  - scipy
-  - mpi4py
-  - h5py
-  - sympy
-  - pytest
-  - ipython
-  - pyvista
-  - pip
-  - pip:
-      - gmsh
-      - gmsh-api
-      - jupyterlab
-      - jupytext
-      - pygments
-      - ipywidgets
-      - trame
-      - trame-vtk
-      - trame-vuetify
-      - pint
