### 2025-01-24 a0c8cdf
Moving as many packages from pip -> conda
