### 2025-01-24 8bd6de9
Align codespaces to the new conda environment
