### 2025-02-06 7c0c632
Create UW3 docker
using micromamba
