+++ environment.yml
@@ -21,6 +21,7 @@ dependencies:
   - trame
   - ipywidgets
   - gmsh
+  - python-gmsh
   - jupyterlab
   - jupytext
   - trame-vtk
@@ -28,5 +29,5 @@ dependencies:
   - pygments
   - pint
   - pip
-  - pip:
-    - gmsh-api # is this used still in UW?
+  # - pip:
+  #   - gmsh-api # is this used still in UW?
