### 2025-02-06 a6b001c
Unify conda environments (#289)
* Attempt to align the 2 conda environments.

More testing to come.

* more tweaks to environment.yaml

* Unifying conda environments for mac and intel

* Removing `environment-macos.yml`
* Adding Lior's packages
* Comment on memory leak in parallel test (to revisit)

* Moving as many packages from pip -> conda

* Align codespaces to the new conda environment

* Create UW3 docker

using micromamba

---------

Co-authored-by: Ben Knight <55677727+bknight1@users.noreply.github.com>
