+++ setup.cfg
@@ -16,7 +16,7 @@ install_requires =
   typeguard
   petsc4py >=3.20
   mpi4py
-  gmsh
+  # gmsh ### does not work on linux arm architecture but can be installed through conda instead
   pint
   psutil
   typing_extensions
