+++ environment.yml
@@ -4,32 +4,30 @@ channels:
 dependencies:
   - python=3.11
   - compilers
-  - openmpi
-  - openmpi-mpicc
-  - openmpi-mpicxx
-  - petsc=3.21.5 # Fix for now (errors with hdf5 / uw3)
+  - mpich
+  - petsc=3.21.5
   - petsc4py=3.21.5
   - cython=3.*
   - mpmath<=1.3
   - mesalib
   - pyvista
-  # gmsh dependencies (prefer not to use apt.txt)
-  - xorg-libxft
-  - xorg-libxinerama
   - pytest
   - numpy<2
+  - scipy
+  - sympy
+  - ipython
+  - mpi4py
+  - h5py=*=mpi*
+  - trame
+  - ipywidgets
+  - gmsh
+  - python-gmsh
+  - jupyterlab
+  - jupytext
+  - trame-vtk
+  - trame-vuetify
+  - pygments
+  - pint
   - pip
-  - pip:
-    - scipy
-    - sympy
-    - ipython
-    - mpi4py
-    - h5py
-    - gmsh
-    - gmsh-api
-    - jupyterlab
-    - jupytext
-    - pygments
-    - trame-vtk
-    - trame-vuetify
-    - pint
+  # - pip:
+  #   - gmsh-api # is this used still in UW?
