### 2025-02-12 e41a918
Install script updates
