+++ setup.cfg
@@ -9,7 +9,7 @@ name = underworld3
 zip_safe = True
 install_requires = 
   xxhash
-  numpy <2
+  numpy
   scipy
   cython
   sympy >=1.12
