+++ environment.yml
@@ -22,12 +22,14 @@ dependencies:
   - ipywidgets
   - gmsh
   - python-gmsh
+  - python-xxhash
   - jupyterlab
   - jupytext
   - trame-vtk
   - trame-vuetify
   - pygments
   - pint
+  - typeguard
   - pip
   # - pip:
   #   - gmsh-api # is this used still in UW?
