### 2025-02-12 25235f5
Remove NPY warning by adding macros definition
