### 2025-02-12 8ffaa6d
Adding new package compilation bits.
* Adding pyproject.toml for PEP518-ness. Now we can explore this more.
* Adding a compile.sh. With pyproject.toml we always need
  --no-build-isolation on
  * compile.sh always outputs to `compile.log`
  * `compile.log` is gitignored
