+++ compile.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+set -x  # show commands
+
+pip install . --no-build-isolation -v | tee compile.log
