### 2025-02-12 dbcbbde
update github workflow to use compile.sh
