### 2025-02-12 e8b1e36
Adding documentation.
* Section on installation is update.
* Added a section on uninstallation.
* Added a `pip install -e` to build_api_docs.sh - that must be done
  otherwise pdoc fails.
