### 2025-02-13 f855236
Disable verbose compile and remove git clean from clean.sh
