### 2025-02-20 #7831 @diehlpk
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> ✅ OK DOIs
>
> - 10.2172/2337606 is OK
> - 10.21105/joss.01136 is OK
> - 10.1016/j.advwatres.2011.04.013 is OK
> - 10.1109/MCSE.2021.3059263 is OK
> - 10.21105/joss.01797 is OK
> - 10.7717/peerj-cs.103 is OK
> - 10.1016/j.pepi.2007.06.009 is OK
> - 10.1016/S0021-9991(02)00031-1 is OK
>
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Achieving High Performance with Unified Residual E...
>
> ❌ MISSING DOIs
>
> - 10.1109/mcse.2010.118 may be a valid DOI for title: Cython: The Best of Both Worlds
>
> ❌ INVALID DOIs
>
> - None
> ```

 @lmoresi please have a look at the missing DOI.
