### 2025-02-24 35a037f
plot correct scalar magnitude in plot vector (#298)
Nice jobs!
