### 2025-03-04 35c6b02
Adding triangle distance routines
Geometry helpers - adding a distance-to-triangles routine. With kdtree this can be extended to a distance-to-triangulation but this is on the to-do list as a helper
