### 2025-03-04 f87e574
Darcy projection solver tolerance
In Darcy solver, the projection routine should not set its own tolerance and over-write any user settings
