+++ src/underworld3/systems/solvers.py
@@ -316,8 +316,9 @@ class SNES_Darcy(SNES_Scalar):
 
         # Now solve flow field
 
-        self._v_projector.petsc_options["snes_rtol"] = 1.0e-6
-        self._v_projector.petsc_options.delValue("ksp_monitor")
+        # self._v_projector.petsc_options["snes_rtol"] = 1.0e-6
+        # self._v_projector.petsc_options.delValue("ksp_monitor")
+        self._v_projector.uw_function = self.darcy_flux
         self._v_projector.solve(zero_init_guess)
 
         return
