### 2025-03-04 3dce801
Add vanishingly small quantity expression to the math functions / helpers
