+++ src/underworld3/maths/functions.py
@@ -15,3 +15,7 @@ def delta(
     delta_fn = sympy.exp(-(x**2) / (2 * epsilon**2)) / (epsilon * sqrt_2pi)
 
     return delta_fn
+
+# Definitions of expressions that are used in various places within underworld
+
+vanishing = function.expressions.UWexpression(r"\varepsilon", 1e-18 , "vanishingly small value")
