### 2025-03-06 6d54069
Preliminary implementation of deferred-evaluation derivative expression
