### 2025-03-17 underworldcode/underworld3#302 @gassmoeller
[JOSS review] Licence.md refers to wrong licence file name
The top-level file Licence.md refers to the file LICENCE for the actual license text, however the correct file name would be [LGPLv3.txt](https://github.com/underworldcode/underworld3/blob/main/LGPLv3.txt)
