### 2025-03-17 underworldcode/underworld3#303 @gassmoeller
[JOSS Review] Local pytest run: 1 test fails
**Describe the runtime issue**
When I run the test-suite after installing underworld3 using mamba following the instructions [here](https://underworldcode.github.io/underworld3/development/_quickstart/Installation.html#source-code-build-using-mamba) in a new clean environment all tests pass except for `tests/test_1100_AdvDiffCartesian.py::test_advDiff_boxmesh[mesh1]`. Which fails with the error posted below. It looks like the automatic CI environment is slightly different (using a preinstalled PETSc), so maybe this is expected, but I wanted to ask if someone can reproduce this and if there is something wrong with the test, or if this is expected behavior.

Related to https://github.com/openjournals/joss-reviews/issues/7831.

**To Reproduce**
Install as described above, run testsuite as listed in the screen output below.

**Screenshots**
```
(uw3) rene@rene-laptop:~/work/reviews/2025-joss-underworld3/underworld3$ pytest
======================================================================================================= test session starts ========================================================================================================
rootdir: /home/rene/work/reviews/2025-joss-underworld3/underworld3
configfile: pyproject.toml
plugins: typeguard-4.4.2, anyio-4.9.0
collected 127 items

tests/test_0000_imports.py .........                                                                                                                                                                                         [  7%]
tests/test_0001_meshes.py ................                                                                                                                                                                                   [ 19%]
tests/test_0002_swarm.py ..                                                                                                                                                                                                  [ 21%]
tests/test_0003_save_load.py ....                                                                                                                                                                                            [ 24%]
tests/test_0004_pointwise_fns.py ...                                                                                                                                                                                         [ 26%]
tests/test_0101_kdtree.py ........                                                                                                                                                                                           [ 33%]
tests/test_0501_integrals.py .........x                                                                                                                                                                                      [ 40%]
tests/test_0503_evaluate.py ............                                                                                                                                                                                     [ 50%]
tests/test_0504_projections.py ...                                                                                                                                                                                           [ 52%]
tests/test_0601_mesh_vector_calc.py .........X..................                                                                                                                                                             [ 74%]
tests/test_1000_poissonCart.py ....                                                                                                                                                                                          [ 77%]
tests/test_1001_poissonSph.py ..                                                                                                                                                                                             [ 79%]
tests/test_1004_DarcyCartesian.py ...                                                                                                                                                                                        [ 81%]
tests/test_1010_stokesCart.py ....                                                                                                                                                                                           [ 85%]
tests/test_1011_stokesSph.py ...                                                                                                                                                                                             [ 87%]
tests/test_1050_VEstokesCart.py .....                                                                                                                                                                                        [ 91%]
tests/test_1100_AdvDiffCartesian.py .F.                                                                                                                                                                                      [ 93%]
tests/test_1100_SLVectorCartesian.py ...                                                                                                                                                                                     [ 96%]
tests/test_1110_advDiffAnnulus.py .                                                                                                                                                                                          [ 96%]
tests/test_1120_SLVectorCartesian.py ...                                                                                                                                                                                     [ 99%]
tests/test_1130_IndexSwarmVariable.py .                                                                                                                                                                                      [100%]

============================================================================================================= FAILURES =============================================================================================================
___________________________________________________________________________________________________ test_advDiff_boxmesh[mesh1] ____________________________________________________________________________________________________

mesh = <underworld3.discretisation.Mesh object at 0x71b66d8a5810>

    @pytest.mark.parametrize(
        "mesh",
        [
            meshStructuredQuadBox,
            unstructured_simplex_box_irregular,
            unstructured_simplex_box_regular,
        ],
    )
    def test_advDiff_boxmesh(mesh):
        print(f"Mesh - Coordinates: {mesh.CoordinateSystem.type}")

        # Create an mesh vars
        v = uw.discretisation.MeshVariable("U", mesh, mesh.dim, degree=1)
        T = uw.discretisation.MeshVariable("T", mesh, 1, degree=3)

        # #### Create the advDiff solver

        adv_diff = uw.systems.AdvDiffusion(
            mesh,
            u_Field=T,
            V_fn=v,
        )

        # ### Set up properties of the adv_diff solver
        # - Constitutive model (Diffusivity)
        # - Boundary conditions
        # - Internal velocity
        # - Initial temperature distribution

        adv_diff.constitutive_model = uw.constitutive_models.DiffusionModel
        adv_diff.constitutive_model.Parameters.diffusivity = kappa

        ### fix temp of top and bottom walls
        adv_diff.add_dirichlet_bc(tmin, "Bottom")
        adv_diff.add_dirichlet_bc(tmin, "Top")
        # adv_diff.add_dirichlet_bc(0., "Left")
        # adv_diff.add_dirichlet_bc(0., "Right")

        with mesh.access(v):
            # initialise fields
            # v.data[:,0] = -1*v.coords[:,1]
            v.data[:, 1] = velocity

        with mesh.access(T):
            T.data[...] = tmin

            pipePosition = ((ymax - ymin) - pipe_thickness) / 2.0

            T.data[
                (T.coords[:, 1] >= (T.coords[:, 1].min() + pipePosition))
                & (T.coords[:, 1] <= (T.coords[:, 1].max() - pipePosition))
            ] = tmax

        # ### Create points to sample the UW results
        ### y coords to sample
        sample_y = np.arange(
            mesh.data[:, 1].min(), mesh.data[:, 1].max(), 0.1 * mesh.get_min_radius()
        )  ### Vertical profile

        ### x coords to sample
        sample_x = np.zeros_like(sample_y)  ### LHS of box

        sample_points = np.empty((sample_x.shape[0], 2))
        sample_points[:, 0] = sample_x
        sample_points[:, 1] = sample_y
        # -

        ### get the initial temp profile
        T_orig = uw.function.evaluate(T.sym[0], sample_points)

        #### 1D diffusion function
        #### To compare UW results with a numerical results

        def diffusion_1D(sample_points, T0, diffusivity, vel, time_1D):
            x = sample_points
            T = T0
            k = diffusivity
            time = time_1D

            dx = sample_points[1] - sample_points[0]

            dt_dif = dx**2 / k
            dt_adv = dx / velocity

            dt = 0.5 * min(dt_dif, dt_adv)

            if time > 0:
                """determine number of its"""
                nts = math.ceil(time / dt)

                """ get dt of 1D model """
                final_dt = time / nts

                for i in range(nts):
                    qT = -k * np.diff(T) / dx
                    dTdt = -np.diff(qT) / dx
                    T[1:-1] += dTdt * final_dt

            return T

        model_time = 0.0

        #### Solve
        dt_est = adv_diff.estimate_dt()

        # This should be stable, and soluble by the 1D FD
        dt = 0.001

        ### diffuse through underworld
        adv_diff.solve(timestep=dt)

        model_time += dt

        ### compare UW and 1D numerical solution
        T_UW = uw.function.evalf(T.sym[0], sample_points)

        T_1D_model = diffusion_1D(
            sample_points=sample_points[:, 1],
            T0=T_orig.copy(),
            diffusivity=kappa,
            vel=velocity,
            time_1D=model_time,
        )

        #### 1D numerical advection
        new_y = sample_points[:, 1] + (velocity * model_time)

        ### some issues with the projection of data onto the sample points so high rtol
>       assert np.allclose(T_UW, T_1D_model, atol=0.2)
E       assert False
E        +  where False = <function allclose at 0x71b6bf333370>(array([0.5       , 0.49999457, 0.49999038, 0.49998696, 0.49998391,\n       0.49998082, 0.49997719, 0.49997227, 0.499977...0000719,\n       0.50001634, 0.50001191, 0.50000876, 0.50000655, 0.50000487,\n       0.5000034 , 0.50000191, 0.50000023]), array([0.5       , 0.5       , 0.5       , 0.5       , 0.5       ,\n       0.49999999, 0.49999999, 0.49999999, 0.499999...9999998,\n       0.49999999, 0.49999999, 0.49999999, 0.5       , 0.5       ,\n       0.5       , 0.5       , 0.5       ]), atol=0.2)
E        +    where <function allclose at 0x71b6bf333370> = np.allclose

tests/test_1100_AdvDiffCartesian.py:180: AssertionError
------------------------------------------------------------------------------------------------------- Captured stdout call -------------------------------------------------------------------------------------------------------
Mesh - Coordinates: Cartesian 2D
===================================================================================================== short test summary info ======================================================================================================
FAILED tests/test_1100_AdvDiffCartesian.py::test_advDiff_boxmesh[mesh1] - assert False
================================================================================= 1 failed, 124 passed, 1 xfailed, 1 xpassed in 181.95s (0:03:01) ==================================================================================
WARNING! There are options you set that were not used!
WARNING! could be spelling mistake, etc!
There are 11 unused database options. They are:
Option left: name:-Solver_341_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_351_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_361_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_371_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_411_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_435_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_459_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_483_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-Solver_505_fieldsplit_pressure_pc_gasm_type value: basic source: code
Option left: name:-viewer_hdf5_collective value: false source: code
Option left: name:-viewer_hdf5_sp_output value: true source: code
(uw3) rene@rene-laptop:~/work/reviews/2025-joss-underworld3/underworld3$ git branch
* development
  main
```

**Environment (please complete the following information):**
 - OS: Ubuntu
 - Version 22.04

mamba list:
```
(uw3) rene@rene-laptop:~/work/reviews/2025-joss-underworld3/underworld3$ mamba list
# packages in environment at /home/rene/mambaforge/envs/uw3:
#
# Name                    Version                   Build  Channel
_openmp_mutex             4.5                  3_kmp_llvm    conda-forge
aiohappyeyeballs          2.6.1              pyhd8ed1ab_0    conda-forge
aiohttp                   3.11.13         py311h2dc5d0c_0    conda-forge
aiosignal                 1.3.2              pyhd8ed1ab_0    conda-forge
alsa-lib                  1.2.13               hb9d3cd8_0    conda-forge
anyio                     4.9.0                    pypi_0    pypi
aom                       3.9.1                hac33072_0    conda-forge
argon2-cffi               23.1.0                   pypi_0    pypi
argon2-cffi-bindings      21.2.0                   pypi_0    pypi
arrow                     1.3.0                    pypi_0    pypi
asttokens                 3.0.0              pyhd8ed1ab_1    conda-forge
async-lru                 2.0.5                    pypi_0    pypi
attr                      2.5.1                h166bdaf_1    conda-forge
attrs                     25.3.0             pyh71513ae_0    conda-forge
babel                     2.17.0                   pypi_0    pypi
beautifulsoup4            4.13.3                   pypi_0    pypi
binutils                  2.40                 h4852527_7    conda-forge
binutils_impl_linux-64    2.40                 ha1999f0_7    conda-forge
binutils_linux-64         2.40                 hb3c18ed_4    conda-forge
bleach                    6.2.0                    pypi_0    pypi
blosc                     1.21.6               he440d0b_1    conda-forge
brotli                    1.1.0                hb9d3cd8_2    conda-forge
brotli-bin                1.1.0                hb9d3cd8_2    conda-forge
brotli-python             1.1.0           py311hfdbb021_2    conda-forge
bzip2                     1.0.8                h4bc722e_7    conda-forge
c-ares                    1.34.4               hb9d3cd8_0    conda-forge
c-compiler                1.5.2                h0b41bf4_0    conda-forge
ca-certificates           2025.1.31            hbcca054_0    conda-forge
cached-property           1.5.2                hd8ed1ab_1    conda-forge
cached_property           1.5.2              pyha770c72_1    conda-forge
cairo                     1.18.4               h3394656_0    conda-forge
certifi                   2025.1.31          pyhd8ed1ab_0    conda-forge
cffi                      1.17.1          py311hf29c0ef_0    conda-forge
charset-normalizer        3.4.1              pyhd8ed1ab_0    conda-forge
colorama                  0.4.6              pyhd8ed1ab_1    conda-forge
comm                      0.2.2                    pypi_0    pypi
compilers                 1.5.2                ha770c72_0    conda-forge
contourpy                 1.3.1           py311hd18a35c_0    conda-forge
cpython                   3.11.11         py311hd8ed1ab_2    conda-forge
cxx-compiler              1.5.2                hf52228f_0    conda-forge
cycler                    0.12.1             pyhd8ed1ab_1    conda-forge
cyrus-sasl                2.1.27               h54b06d7_7    conda-forge
cython                    3.0.12          py311ha3e34f5_0    conda-forge
dav1d                     1.2.1                hd590300_0    conda-forge
dbus                      1.13.6               h5008d03_3    conda-forge
debugpy                   1.8.13                   pypi_0    pypi
decorator                 5.2.1              pyhd8ed1ab_0    conda-forge
defusedxml                0.7.1                    pypi_0    pypi
double-conversion         3.3.1                h5888daf_0    conda-forge
elfutils                  0.192                h7f4e02f_1    conda-forge
exceptiongroup            1.2.2              pyhd8ed1ab_1    conda-forge
executing                 2.1.0              pyhd8ed1ab_1    conda-forge
expat                     2.6.4                h5888daf_0    conda-forge
fastjsonschema            2.21.1                   pypi_0    pypi
ffmpeg                    7.1.1           gpl_h332a80e_501    conda-forge
fftw                      3.3.10          mpi_openmpi_h99e62ba_10    conda-forge
flexcache                 0.3                      pypi_0    pypi
flexparser                0.4                      pypi_0    pypi
font-ttf-dejavu-sans-mono 2.37                 hab24e00_0    conda-forge
font-ttf-inconsolata      3.000                h77eed37_0    conda-forge
font-ttf-source-code-pro  2.038                h77eed37_0    conda-forge
font-ttf-ubuntu           0.83                 h77eed37_3    conda-forge
fontconfig                2.15.0               h7e30c49_1    conda-forge
fonts-conda-ecosystem     1                             0    conda-forge
fonts-conda-forge         1                             0    conda-forge
fonttools                 4.56.0          py311h2dc5d0c_0    conda-forge
fortran-compiler          1.5.2                hdb1a99f_0    conda-forge
fqdn                      1.5.1                    pypi_0    pypi
freetype                  2.13.3               h48d6fc4_0    conda-forge
fribidi                   1.0.10               h36c2ea0_0    conda-forge
frozenlist                1.5.0           py311h2dc5d0c_1    conda-forge
gcc                       11.4.0              h602e360_13    conda-forge
gcc_impl_linux-64         11.4.0              h00c12a0_13    conda-forge
gcc_linux-64              11.4.0               ha077dfb_4    conda-forge
gdk-pixbuf                2.42.12              hb9ae30d_0    conda-forge
gettext                   0.23.1               h5888daf_0    conda-forge
gettext-tools             0.23.1               h5888daf_0    conda-forge
gfortran                  11.4.0              h602e360_13    conda-forge
gfortran_impl_linux-64    11.4.0              hd465bd8_13    conda-forge
gfortran_linux-64         11.4.0               hddf7001_4    conda-forge
gl2ps                     1.4.2                hae5d5c5_1    conda-forge
glew                      2.1.0                h9c3ff4c_2    conda-forge
gmp                       6.3.0                hac33072_2    conda-forge
gmpy2                     2.1.5           py311h0f6cedb_3    conda-forge
gmsh                      4.13.1                   pypi_0    pypi
gmsh-api                  4.13.0                   pypi_0    pypi
gnutls                    3.8.9                h5746830_0    conda-forge
graphite2                 1.3.13            h59595ed_1003    conda-forge
gxx                       11.4.0              h602e360_13    conda-forge
[…]
