### 2025-03-17 underworldcode/underworld3#304 @gassmoeller
[JOSS review] API documentation offline?
It looks like the website for the API documentation is offline, at least the link to https://underworldcode.github.io/underworld3/development_api/ which is linked on this page: https://underworldcode.github.io/underworld3/development/_quickstart/NextSteps.html gives me an `404 - page not found` error.

Related to https://github.com/openjournals/joss-reviews/issues/7831
