### 2025-03-17 underworldcode/underworld3#305 @gassmoeller
[JOSS review] Information about contributing
I was looking for information on how to contribute to underworld3 as part of https://github.com/openjournals/joss-reviews/issues/7831 (subsection Documentation - Community Guidelines), but I couldnt find much information. There is neither a `Contributing.md` document in the repository, nor is there any information about contributing in the Readme or on https://underworldcode.github.io/underworld3/development/_quickstart/index.html. I am sure it is probably somewhere on some website that I havent found, but it is not available in the standard places.
