### 2025-03-17 underworldcode/underworld3#306 @gassmoeller
[JOSS review] Paper comments
Here is a list of small comments about https://github.com/underworldcode/underworld3/blob/development/docs/joss-paper/paper.md. Part of https://github.com/openjournals/joss-reviews/issues/7831#issuecomment-2705502693.

- The statement of need could use one sentence at the beginning containing examples of actual applications that Underword3 is designed to solve. Currently you mention the properties of the models (nonlinear rheology, strain weakening, ...), but applications like mantle convection, lithosphere dynamics, porous flow, crustal deformation would make it much easier to users to understand if the code would work for them.
- JOSS requires a section on the state of the field and how Underworld3 fits into it. This is currently missing. From the JOSS guidelines: `State of the field: Do the authors describe how this software compares to other commonly-used packages?` I think you already describe in depth what makes Underworld special. One additional sentence would be sufficient to mention that there are packages fully based on python and matlab with examples, and fully compiled modeling packages with some examples, and that Underworld bridges the gap (or something similar, the content is up to you, but you need that section/comparison).
- The section on the mathematical framework is pretty dense with little introduction and explanation. I know this was probably done to save space and keep a 2-page limit, but in order to be more useful I would suggest to at least:
  - rephrase the sentence at the beginning to something like this: `Underworld generates the weak forms of the equations through PETSc, which provides a template form for the automatic generation ...
  - make sure in line 51 and 52 to mention which symbols in the equation represent the fluxes (F), forces (f) and unknowns (u).
  - equations 2 and 3 contain a number of unexplained symbols that would be necessary to spell out to make this useful to the average reader.
  - add one sentence at the end of the section that explains what the benefit of using this framework is in the context of underworld.

Generally I think the paper is in great shape, and with the small changes above (even if it adds half a page of length or so) should be ready to be published.
