### 2025-03-20 ceac2d9
Fixes to anisotropic models and triangulation stuff
