### 2025-03-26 underworldcode/underworld3#303 @lmoresi
It's difficult to have a reliable test for advection because of the potential for small phase errors to offset the solution in space when we'd probably judge it as OK by eye (given low resolution for tests). I'll have a think about a different metric for this test.
