### 2025-03-26 underworldcode/underworld3#303 @bknight1
I can update this one. I have a slightly different benchmark for my diffusion paper that uses an analytical solution and starts from a partially diffused state.
