+++ test.sh
@@ -22,5 +22,5 @@ $PYTEST tests/test_100[0-9]*py
 $PYTEST tests/test_1010*py tests/test_1011*py tests/test_1050*py
 
 # Diffusion / Advection tests
-# $PYTEST tests/test_1100*py
-# $PYTEST tests/test_1110*py # Annulus version
+$PYTEST tests/test_1100*py
+$PYTEST tests/test_1110*py # Annulus version
