### 2025-03-26 underworldcode/underworld3#303 @lmoresi
Thanks, @bknight1. I had commented out the two advection tests (they work for me but not for everyone)  but it would be good to get a replacement that is more robust.
