### 2025-03-26 underworldcode/underworld3#306 @lmoresi
@gassmoeller - thanks, that's very helpful.

Yes, the density does come from the relatively short length recommendation but a more readable paper will probably be quicker to read than a dense one. I'll put my mind to it.
