### 2025-03-26 underworldcode/underworld3#304 @lmoresi
We moved the documentation to a sub-directory as part of a reorganisation post-submission and I forgot that the joss-submission branch does not have it's own API docs build. I've updated the link but, note that the documentation is for main (mostly unchanged while we go through this review) and development (changes monthly).
