+++ src/underworld3/__init__.py
@@ -12,17 +12,17 @@ a notebook.
 
 ## Documentation
 
-The underworld documentation is in two parts: the user manual / theory manual is a hand-written document that is built from this repository automatically 
-from the sources in the `Jupyterbook` directory. This api documentation is autogenerated from the code and its annotations. There are two version of 
+The underworld documentation is in two parts: the user manual / theory manual is a hand-written document that is built from this repository automatically
+from the sources in the `Jupyterbook` directory. This api documentation is autogenerated from the code and its annotations. There are two version of
 each documentation package that are published online. The main (stable) branch
 
-- https://underworldcode.github.io/underworld3/main/FrontPage.html
-- https://underworldcode.github.io/underworld3/main_api/index.html
+- https://underworldcode.github.io/underworld3/main/index.html
+- https://underworldcode.github.io/underworld3/main_api/underworld3/index.html
 
 The development branch has similar documentation:
 
-- https://underworldcode.github.io/underworld3/development/FrontPage.html
-- https://underworldcode.github.io/underworld3/development_api/index.html
+- https://underworldcode.github.io/underworld3/development/index.html
+- https://underworldcode.github.io/underworld3/development_api/underworld3/index.html
 
 
 ## Building / installation
