+++ docs/user/NextSteps.qmd
@@ -16,7 +16,7 @@ In addition to the notebooks in this brief set of examples, there are a number o
 
   - [The Underworld Website / Blog](https://www.underworldcode.org)
 
-  - [The API documentation](https://underworldcode.github.io/underworld3/development_api/index.html)
+  - [The API documentation](https://underworldcode.github.io/underworld3/development_api/underworld3/index.html)
   (all the modules and functions and their full sets of arguments) is automatically generated from the source code and uses the same rich markdown content as the notebook help text.
 
   - The [`underworld3` GitHub repository](https://github.com/underworldcode/underworld3) is the most active development community for the code.
