### 2025-03-27 43d77fd
Update NextSteps.qmd
Really fix the links
