+++ docs/user/UW-Background.bib
@@ -18,6 +18,7 @@
   volume = {13},
   number = {2},
   pages = {31--39},
+  doi = {10.1109/mcse.2010.118},
   publisher = {IEEE}
 }
 
