### 2025-03-28 ff66538
Update build_deploy_pdoc.yml
Fixing ubuntu version
