### 2025-03-28 af61e0e
Update advDiff test
Fix for #303
