### 2025-03-30 underworldcode/underworld3#304 @lmoresi
I fixed the CI issues (out of date ubuntu in GH actions).
I fixed the links to the api in the README and docs of the branches that generate documentation.
