### 2025-04-08 #7831 @diehlpk
Hi @lmoresi how is the progress on the open issues?
