### 2025-04-08 e2893fa
Fix vector magnitude - sqrt was missing
