+++ docs/joss-paper/paper.bib
@@ -1,4 +1,4 @@
-@techreport{balay.etal.PETSc.2024,
+@techreport{balayPETScTAOUsers2024,
   title = {{{PETSc}}/{{TAO Users Manual V}}.3.21},
   author = {Balay, S. and Abhyankar, S. and Adams, M. and Benson, S. and Brown, J. and Brune, P. and Buschelman, K. and Constantinescu, E. and Dalcin, L. and Dener, A. and Eijkhout, V. and Faibussowitsch, J. and Gropp, W. and Hapla, V. and Isaac, T. and Jolivet, P. and Karpeev, D. and Kaushik, D. and Knepley, M. and Kong, F. and Kruger, S. and May, D. and McInnes, L. and Mills, R. and Mitchell, L. and Munson, T. and Roman, J. and Rupp, K. and Sanan, P. and Sarich, J. and Smith, B. and Zampini, S. and Zhang, H. and Zhang, H. and Zhang, J.},
   year = {2024},
@@ -10,6 +10,22 @@
   file = {/Users/lmoresi/Zotero/storage/FTX49W6P/Balay et al. - 2024 - PETScTAO Users Manual V.3.21.pdf}
 }
 
+@incollection{batheFiniteElementMethod2008,
+  title = {Finite {{Element Method}}},
+  booktitle = {Wiley {{Encyclopedia}} of {{Computer Science}} and {{Engineering}}},
+  author = {Bathe, Klaus-J{\dbend}{\dbend}rgen},
+  editor = {Wah, Benjamin W.},
+  year = {2008},
+  month = jun,
+  pages = {ecse159},
+  publisher = {John Wiley \& Sons, Inc.},
+  address = {Hoboken, NJ, USA},
+  doi = {10.1002/9780470050118.ecse159},
+  urldate = {2022-04-27},
+  isbn = {978-0-470-05011-8},
+  langid = {english}
+}
+
 @article{behnel2011cython,
   title = {Cython: {{The}} Best of Both Worlds},
   author = {Behnel, Stefan and Bradshaw, Robert and Citro, Craig and Dalcin, Lisandro and Seljebotn, Dag Sverre and Smith, Kurt},
@@ -18,11 +34,11 @@
   volume = {13},
   number = {2},
   pages = {31--39},
-  doi = {10.1109/mcse.2010.118},
-  publisher = {IEEE}
+  publisher = {IEEE},
+  doi = {10.1109/mcse.2010.118}
 }
 
-@article{beucher.etal.UWGeodynamics.2019,
+@article{beucherUWGeodynamicsTeachingResearch2019,
   title = {{{UWGeodynamics}}: {{A}} Teaching and Research Tool for Numerical Geodynamic Modelling},
   shorttitle = {{{UWGeodynamics}}},
   author = {Beucher, Romain and Moresi, Louis and Giordani, Julian and Mansour, John and Sandiford, Dan and Farrington, Rebecca and Mondy, Luke and Mallard, Claire and Rey, Patrice and Duclaux, Guillaume and Kaluza, Owen and Laik, Arijit and Mor{\'o}n, Sara},
@@ -39,6 +55,38 @@ ZSCC: NoCitationData[s0]},
   file = {/Users/lmoresi/Zotero/storage/G5UVAHXM/Beucher et al. - 2019 - UWGeodynamics A teaching and research tool for nu.pdf}
 }
 
+@misc{boriskausLaMEM2024,
+  title = {{{LaMEM}}},
+  shorttitle = {{{UniMainzGeo}}/{{LaMEM}}},
[…]
