+++ docs/joss-paper/paper.md
@@ -115,8 +115,8 @@ The symbolic representation of the strong-form that is encoded in `underworld3` 
 
 \begin{equation}\label{eq:sympy-strong-form}
  \Bigl[ {D u}/{D t} \Bigr]
--\nabla \cdot \Bigl[ \sigma(u, \nabla u, \mathbf{x}, t)) \Bigr]
--\Bigl[ \mathrm{h}(u, \nabla u, \mathbf{x},t) \Bigr]
+-\nabla \cdot \Bigl[ \sigma(u, \nabla u, \mathbf{x}, t) \Bigr]
+-\Bigl[ \mathrm{H}(u, \nabla u, \mathbf{x},t) \Bigr]
 = 0
 \end{equation}
 
@@ -128,10 +128,9 @@ The main benefits of combining `sympy` with the `PETSc` weak form template is a 
 
 # State of the Field
 
-`Underworld3` is one among a small number of specialised codes for studying Earth deformation on medium to long geological time-scales. Early geodynamics codes, of which there were too many to recite individually, were highly specialised for specific tasks. A second generation of codes are built around generic partial differential equation solvers with scriptable interfaces.  These include: Aspect [C++ plugin architecture: @heisterHighAccuracyMantle2017], Underworld 1 and 2 [xml object composition / python scripting respectively, @mansourUnderworld2PythonGeodynamics2020; @moresiComputationalApproachesStudying2007], Fluidity [xml combined with python scripting, @daviesFluidityFullyUnstructured2011a], Milamin [Matlab front end, @dabrowskiMILAMINMATLABbasedFinite2008], LaMEM [julia scripting @boriskausLaMEM2024], TerraFERMA [Unified Form Language, @wilsonTerraFERMARansparent2017], GAdopt [Unified Form Language / python @daviesAutomaticFiniteelementMethods2022].
+`Underworld3` is one among a small number of codes for studying Earth deformation on medium to long geological time-scales. Early geodynamics codes, of which there were too many to recite individually, were highly specialised for specific tasks with little flexibility for user-defined problems. A subsequent generation of codes, currently in use, was built around generic partial differential equation solvers with scriptable interfaces.  These include: Aspect [C++ plugin architecture: @heisterHighAccuracyMantle2017], Underworld 1 and 2 [xml object composition / python scripting respectively, @mansourUnderworld2PythonGeodynamics2020; @moresiComputationalApproachesStudying2007], Fluidity [xml combined with python scripting, @daviesFluidityFullyUnstructured2011a], Milamin [Matlab front end, @dabrowskiMILAMINMATLABbasedFinite2008], LaMEM [julia scripting @boriskausLaMEM2024], TerraFERMA [Unified Form Language, @wilsonTerraFERMARansparent2017], GAdopt [Unified Form Language / python @daviesAutomaticFiniteelementMethods2022].
 
-`Underworld3` uses python and the python package `sympy` as its scripting interface. The advantage of `sympy` is that it is a fully featured symbolic algebra package which allows much of the logic of the mathematical problem description to be defined symbolically and dynamically rather than as static relationships between objects. It also provides deep, mathematical introspection when developing and debugging models.
-
+`Underworld3` uses python and the python package `sympy` as the scripting interface that overlies the generic partial differential equation layer. The advantage of `sympy` is that it is a fully featured symbolic algebra package which allows much of the logic of the mathematical problem description to be defined symbolically and dynamically rather than as static relationships between objects. It also provides deep, mathematical introspection when developing and debugging models.
 
 # Discussion
 
@@ -150,23 +149,3 @@ The development and testing of our codes is also supported by computational reso
[…]
