### 2025-04-17 #7831 @lmoresi
More-or-less done. I just need to link the specific commits to the issues raised.
