### 2025-04-22 underworldcode/underworld3#306 @lmoresi
@gassmoeller — see comments below. There are a number of commits related to this so I will not point out every single one but  address them here:

~~The statement of need could use one sentence at the beginning containing examples of actual applications that Underword3 is designed to solve. Currently you mention the properties of the models (nonlinear rheology, strain weakening, ...), but applications like mantle convection, lithosphere dynamics, porous flow, crustal deformation would make it much easier to users to understand if the code would work for them.~~

Addressed this by adding the required additional sentence before launching into the detailed discussion which is now in the (shortened) second sentence.

~~JOSS requires a section on the state of the field and how Underworld3 fits into it. This is currently missing. From the JOSS guidelines: State of the field: Do the authors describe how this software compares to other commonly-used packages? I think you already describe in depth what makes Underworld special. One additional sentence would be sufficient to mention that there are packages fully based on python and matlab with examples, and fully compiled modeling packages with some examples, and that Underworld bridges the gap (or something similar, the content is up to you, but you need that section/comparison).~~

I added this section with a number of comparable codes. I limit myself to those that have some sort of scripting / plugin architecture to allow flexible model development. I have shied away from adding completely general modelling codes though since they tend to be set up with a different audience in mind.

~~The section on the mathematical framework is pretty dense with little introduction and explanation. I know this was probably done to save space and keep a 2-page limit, but in order to be more useful I would suggest to at least:
rephrase the sentence at the beginning to something like this: `Underworld generates the weak forms of the equations through PETSc, which provides a template form for the automatic generation ...
make sure in line 51 and 52 to mention which symbols in the equation represent the fluxes (F), forces (f) and unknowns (u).
equations 2 and 3 contain a number of unexplained symbols that would be necessary to spell out to make this useful to the average reader.
add one sentence at the end of the section that explains what the benefit of using this framework is in the context of underworld.~~

I have tried to be more careful with definitions and include more explanation. We have blown out the paper length limit already and the additional information in this section adds length but improves readability considerably.

@diehlpk — how do you feel about this ?
