### 2025-04-22 #7831 @lmoresi
@gassmoeller — here are the responses to your review comments:

Issues:

  1. https://github.com/underworldcode/underworld3/issues/302 - fixed
  2. https://github.com/underworldcode/underworld3/issues/303  fixed in commit: https://github.com/underworldcode/underworld3/commit/af61e0eeff72d3a9c04dd5d228b1166035fb40bc
  3. https://github.com/underworldcode/underworld3/issues/304 fixed (several commits) - links to API docs were broken after we reorganised the main / dev branches but did not update the README files in main/dev or joss-submission branches correctly.
  4. https://github.com/underworldcode/underworld3/issues/305: fixed in commit https://github.com/underworldcode/underworld3/commit/73cbaead6238cd6d6e9141b6d9a50ff49394df33
  5. https://github.com/underworldcode/underworld3/issues/306: updated paper.md to reflect comments from @gassmoeller. Specifically: Expanding the mathematics section to make it flow and to ensure all symbols are referenced / described correctly; Adding the missing "State of the Field" section

Checklist Items that are not issues:

  1. Automatic testing: the pytest suite is run for all PRs and also for the API documentation builds on main / dev prior to deployment. I have added a badge in the README which reflects the status of these tests.

This part of the response is complete. I am happy to have @gassmoeller close issues or request further changes.

L
