### 2025-04-24 underworldcode/underworld3#303 @gassmoeller
Thanks for the changes! I know how hard it can be to find reliable tests for processes that can create small oscillations. Having a reference platform like your github action is usually good enough, but your new test also passes on my system now. So this is solved :+1:
