### 2025-04-24 #7831 @diehlpk
@lmoresi Why do you have to `LICENSE.md` and `LICENSE.md`? One has content and one is empty?
