+++ docs/joss-paper/paper.bib
@@ -177,6 +177,7 @@ ZSCC: NoCitationData[s0]},
   edition = {1. Dr.},
   publisher = {Prentice Hall},
   address = {Englewood Cliffs, N.J.},
+  doi = {10.1016/0045-7825(87)90013-2},
   isbn = {978-0-13-317017-7 978-0-13-317025-2},
   langid = {english}
 }
@@ -349,8 +350,8 @@ ZSCC: 0000003},
   edition = {Seventh edition},
   publisher = {Elsevier, Butterworth-Heinemann},
   address = {Amsterdam},
+  doi = {10.1016/b978-1-85617-633-0.00019-7},
   isbn = {978-1-85617-633-0},
   lccn = {TA640.2 .Z5 2013},
-  keywords = {Engineering mathematics,Finite element method,Fluid dynamics},
-  annotation = {OCLC: ocn852808496}
+  keywords = {Engineering mathematics,Finite element method,Fluid dynamics}
 }
