### 2025-05-01 #7831 @lmoresi
> Why do you have to LICENSE.md and LICENSE.md? One has content and one is empty?

@diehlpk LICENSE.md is a symbolic link to the LICENCE.md file. We consistently use the British / Australian spelling of license (verb) / licence (noun) in our publications and documentation but GitHub thinks the code is unlicenced if the LICENSE file is missing - or at least it did when we first started uw3.

> ❌ MISSING DOIs
> - 10.1016/0045-7825(87)90013-2 may be a valid DOI for title: The Finite Element Method: Linear Static and Dynam...
> - 10.1016/b978-1

doi/s missin for books (zenodo issue) - Fixed in commit e5b5dca9aeeaed0912328554de3f37e60e103d23
