### 2025-05-05 554f514
PostHog telemetry (#312)
* Using postHog service to track user telemetry

* Now a unique build key is made everytime the code is build (setup.py
changes)
* That key is in gitignore and clean.sh.
* When underworld3 is imported (__init__py) defines a function only on
  the root proc to send telemetry of the install to postHog service.
  (utilities/_utils/postHog).
* This system can be turned off with the definition of the environment
  variable 'UW_NO_USAGE_METRICS'. This is what's done in test.sh.

* Update src/underworld3/utilities/_utils.py

typo in comment.

Co-authored-by: Copilot <17572847
[…]
