+++ clean.sh
@@ -9,4 +9,5 @@ find . -name __pycache__ -exec rm -r {} +
 rm -rf underworld3.egg-info
 rm -rf .pytest_cache
 rm -rf uw3_api_docs
+rm ./src/underworld3/_uwid.py
 #git clean -dfX
