+++ src/underworld3/utilities/_utils.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 import underworld3 as uw
 import numpy as np
 import io
@@ -235,3 +233,60 @@ def gather_data(val, bcast=False, dtype="float64"):
     comm.barrier()
 
     return val_global
+
+import requests
+import json
+
+# define POSTHOG service bits
+POSTHOG_API_KEY = 'phc_dem097q3pv44QFMK9Fo0sIo96cJPWb9OrSb1FZAowAM'
+POSTHOG_PROJECT_URL = 'https://eu.i.posthog.com'
+
+def postHog( event_name, ev_dict ):
+    """ 
+    Posts an Event Tracking message to PostHog.
+    
+    Current Underworld3 only dispatches an event when the underworld module
+    is imported. This is effected by the calling of this function from
+    underworld/__init__.py.
+    
+    PostHog uses distinct_id in ev_dict to determine unique users. In Underworld this is 
+    generated by a random string at install and record it in _uwid.py (the
+    value is available via the uw._id attribute). If the file (_uwid.py) exists, it
+    is not recreated, so generally it will only be created the first time you build
+    underworld. As this is a 'per build' identifier, it means that all users of a
+    particular docker image will be identified as the same GA user. Likewise, all
+    users of a particular HPC Underworld module will also be identified as the same
+    user.     
+    Regarding HPC usage, it seems that the compute nodes on most machines are closed
+    to external network access, and hence POST requests will not be dispatched
+    successfully. Unfortunately this means that most high proc count simulations
+    will not be captured in this data.
+    
+    This 'trojan' is via PostHog service. It can be disabled by setting the environment 
+    variable `UW_NO_USAGE_METRICS`. Note, this function will return quietly on any errors. 
+    
+    Parameters
+    ----------
+    event_name: str
+        Textual name for event_name. Can only contain alpha-numeric characters and underscores.
+    ev_dict: dict
+        Optional parameter dictionary for event.
+
+    """
+    try:
+        # build payload for PostHog comms
+        payload = {
+            "api_key": POSTHOG_API_KEY,
+            "event": event_name,
+        }
+        payload.update( ev_dict )
+
+        url = f"{POSTHOG_PROJECT_URL}/capture/"
+        headers = {"Content-Type": "application/json"}
+        r = requests.post(url, json=payload, headers=headers, timeout=10)
+
+        # # debugging
+        # print(f"The would-be payload:\n{payload}\n")
+        # print(f"request.post status: {r.status_code}")
+    except Exception as e:
+        print(f"PostHog telemetry failed: {e}")
