+++ src/underworld3/__init__.py
@@ -129,6 +129,53 @@ import underworld3.scaling
 import underworld3.visualisation
 import numpy as _np
 
+# get the id for this installation of uw, if not there, create it
+try:
+    from ._uwid import uwid as _id
+except:
+    import uuid as _uuid
+    _id = str(_uuid.uuid4())
+
+###########################################
+####### For User telemetry metrics ########
+
+# setup metrics function *only* if we are rank==0. 
+# Metric can be disables by setting UW_NO_USAGE_METRICS env var.
+# TODO: confirm not other condition like tests?
+if (underworld3.mpi.rank == 0):
+    def _sendData():
+
+        import os
+        # disable collection of data if requested
+        if "UW_NO_USAGE_METRICS" not in os.environ:
+            # get platform info
+            import platform
+            sysinfo  =        platform.system()
+            sysinfo += "__" + platform.machine()
+            # check if docker
+            if (os.path.isfile("/.dockerinit")):
+                sysinfo += "__docker"
+
+            event_dict = { "properties": {
+                               "version"  : underworld3.__version__,
+                               "platform" : sysinfo,
+                               "run_size" : underworld3.mpi.size,
+                               "distinct_id" : _id, 
+                            }
+                         }
+
+            # send info async
+            import threading
+            thread = threading.Thread( target=underworld3.utilities._utils.postHog, args=("import_uw3", event_dict) )
+            thread.daemon = True
+            thread.start()
+
+    try:
+        _sendData()
+    except: # continue quietly if something above failed
+        pass
+####### END User telemetry metrics ########
+
 # Info for JIT modules.
 # These dicts should be populated by submodules
 # which define cython/c based classes.
