### 2025-05-05 0f1dbfc
Upd (#313)
* Coordinates - 0 dependency bug

When a coordinate system independent of one of the Cartesian basis vectors (e.g. the longitudinal coordinate does not contain z), there is a compilation error in the JIT compiler which complains when terms cancel completely. I have introduced an expression 0(x,y,z) to represent this situation which is epsilon * (x+y+z) as a workaround.

* Bug in Regional Box Mesh boundaries

Numbering error

* Adding triangle distance routines

Geometry helpers - adding a distance-to-triangles routine. With kdtree this can be extended to a distance-to-triangulation 
[…]
