+++ src/underworld3/function/expressions.py
@@ -1,5 +1,6 @@
 import sympy
 from sympy import Symbol, simplify, Number
+import underworld3 as uw
 from underworld3.utilities._api_tools import uw_object
 from underworld3.discretisation import _MeshVariable
 
@@ -49,6 +50,7 @@ def _substitute_one_expr(fn, sub_expr, keep_constants=True, return_self=True):
 # Not sure the best name for this
 def substitute(fn, keep_constants=True, return_self=True):
     return unwrap(fn, keep_constants, return_self)
+
 
 def _unwrap_expressions(fn, keep_constants=True, return_self=True):
     expr = fn
@@ -191,7 +193,7 @@ class UWexpression(uw_object, Symbol):
             )
 
         self.symbol = self._unique_name
-        self.sym = sym     # Accept anything, sympify is opinionated
+        self.sym = sym  # Accept anything, sympify is opinionated
         self.description = description
 
         # this is not being honoured by sympy Symbol so do it by hand
@@ -281,7 +283,7 @@ class UWexpression(uw_object, Symbol):
         return self_s
 
     def dependencies(self, keep_constants=True):
-        return extract_expressions(self, keep_constants)
+        return extract_expressions(self)
 
     def all_dependencies(self, keep_constants=True):
         return extract_expressions_and_functions(self)
@@ -306,12 +308,7 @@ class UWexpression(uw_object, Symbol):
         if sympy.sympify(self.sym) is not None:
             display(
                 Latex(
-                    r"$"
-                    + r"\quad" * level
-                    + "$"
-                    + self._repr_latex_()
-                    + "$=$"
-                    + latex
+                    r"$" + r"\quad" * level + "$" + self._repr_latex_() + "$=$" + latex
                 ),
             )
             if description == True:
@@ -336,3 +333,78 @@ class UWexpression(uw_object, Symbol):
             pass
 
         return
+
+
+class UWDerivativeExpression(UWexpression):
+    """
+    underworld `expressions` are sympy symbols with attached
+    numeric/expression values that are substituted into an underworld function
+    before evaluation.
+
+    derivative expressions are unevaluated / symbolic derivatives that remain
+    symbolic until they need to be evaluated.
+
+    Note - this class would usually be automatically generated by asking for the
+    derivative of an expression with `evaluate=False`
+
+    ```{python}
+        alpha = UWDerivativeExpression(
+                        r'\\alpha',
+                        expr=uw_expression,
+                        diff_expr=diff_expression,
+                        description=fr"\partial{expr.description}/\partial{diff_expr.description}"
+                            )
+        print(alpha.sym)
+        print(alpha.description)
+    ```
+
+    """
+
+    def __init__(
+        self,
+        name,
+        expr,
+        diff_variable,
+        description="derivative of expression provided",
+    ):
+
+        self.symbol = self._unique_name
+        self._sym = None
+        self.diff_variable = None
+
+        self._sym = expr  # Accept anything, sympify is overly opinionated if we try to `sympify`
+        self._diff_variable = diff_variable
+        self.description = description
+
+        # this is not being honoured by sympy Symbol so do it by hand
+        self._uw_id = uw_object._obj_count
+        uw_object._obj_count += 1
+
+        return
+
+    @property
+    def sym(self):
+        return uw.function.derivative(self._sym, self.diff_variable)
+
+    @property
+    def expr(self):
+        return self._sym
+
+    @property
+    def diff_variable(self):
+        return self._diff_variable
+
+    @diff_variable.setter
+    def diff_variable(self, value):
+        self._diff_variable = value
+
+    # TODO: DEPRECATION
+    # The value attribute is no longer needed in the offical release of Underworld3
+    @property
+    def value(self):
[…]
