### 2025-05-05 aa3e853
Update build_deploy_pdoc.yml (#314)
adding pdocs3 to the github action
