### 2025-05-07 41665b7
Changing the behaviour of expressions to ensure:
   1) Symbols have unique names
   2) Where name clashes occur, no expression is built unless
   3) A unique identifier is explicitly requested

UW expressions are essentially immutable containers - only their
contents can change but not the name or description.

Option 3 is intended for private use within solvers where the expression
is not intended to be used other than within the object itself. This
allows us to re-use the symbol for (say) DuDt within duplicate solvers
while keeping consistent with the requirements of the JIT.

The copy metho
[…]
