+++ src/underworld3/cython/petsc_generic_snes_solvers.pyx
@@ -10,8 +10,13 @@ import underworld3
 import underworld3 as uw
 from   underworld3.utilities._jitextension import getext
 import underworld3.timing as timing
+
 from underworld3.utilities._api_tools import uw_object
 from underworld3.utilities._api_tools import class_or_instance_method
+
+from underworld3.function import expression as public_expression
+expression = lambda *x, **X: public_expression(*x, _unique_name_generation=True, **X)
+
 
 include "petsc_extras.pxi"
 
@@ -674,8 +679,8 @@ class SNES_Scalar(SolverBaseClass):
         # f0  = sympy.Array(uw.function.fn_substitute_expressions(self.F0.sym)).reshape(1).as_immutable()
         # F1  = sympy.Array(uw.function.fn_substitute_expressions(self.F1.sym)).reshape(dim).as_immutable()
 
-        f0  = sympy.Array(uw.function.expressions.unwrap(self.F0.sym, keep_constants=False, return_self=False)).reshape(1).as_immutable()
-        F1  = sympy.Array(uw.function.expressions.unwrap(self.F1.sym, keep_constants=False, return_self=False)).reshape(dim).as_immutable()
+        f0  = sympy.Array(uw.function.expression.unwrap(self.F0.sym, keep_constants=False, return_self=False)).reshape(1).as_immutable()
+        F1  = sympy.Array(uw.function.expression.unwrap(self.F1.sym, keep_constants=False, return_self=False)).reshape(dim).as_immutable()
 
         self._u_f0 = f0
         self._u_F1 = F1
@@ -1273,8 +1278,8 @@ class SNES_Vector(SolverBaseClass):
         # f0  = sympy.Array(uw.function.fn_substitute_expressions(self.F0.sym)).reshape(dim).as_immutable()
         # F1  = sympy.Array(uw.function.fn_substitute_expressions(self.F1.sym)).reshape(dim,dim).as_immutable()
 
-        f0  = sympy.Array(uw.function.expressions.unwrap(self.F0.sym, keep_constants=False, return_self=False)).reshape(dim).as_immutable()
-        F1  = sympy.Array(uw.function.expressions.unwrap(self.F1.sym, keep_constants=False, return_self=False)).reshape(dim,dim).as_immutable()
+        f0  = sympy.Array(uw.function.expression.unwrap(self.F0.sym, keep_constants=False, return_self=False)).reshape(dim).as_immutable()
+        F1  = sympy.Array(uw.function.expression.unwrap(self.F1.sym, keep_constants=False, return_self=False)).reshape(dim,dim).as_immutable()
 
 
         self._u_f0 = f0
@@ -1773,7 +1778,7 @@ class SNES_Stokes_SaddlePt(SolverBaseClass):
         self.petsc_options[f"fieldsplit_{p_name}_ksp_type"] = "fgmres"
         self.petsc_options[f"fieldsplit_{p_name}_ksp_rtol"]  = self._tolerance
         self.petsc_options[f"fieldsplit_{p_name}_pc_type"] = "gasm"
-        self.petsc_options[f"fieldsplit_{p_name}_pc_gasm_type"] = "basic"
+        # self.petsc_options[f"fieldsplit_{p_name}_pc_gasm_type"] = "basic"
 
         ## may be more robust but usually slower
         # self.petsc_options[f"fieldsplit_{p_name}_ksp_type"] = "fgmres"
@@ -2106,9 +2111,9 @@ class SNES_Stokes_SaddlePt(SolverBaseClass):
         ## and do these one by one as required by PETSc. However, at the moment, this
         ## is working .. so be careful !!
 
[…]

[…]
