### 2025-05-11 #7831 @lmoresi
@knepley - can you provide a comment on the mathematical framework issues raised by @chennachaos - specifically about the potential problems with the terminology in your paper on the PETSc point wise functions / weak form. The current revision of our paper in pdf form is in the repo here: https://github.com/underworldcode/underworld3/blob/joss-submission/docs/joss-paper/paper.pdf and the equation under suspicion is (2).

@knepley - I think the fact that F(u) is in both the strong and weak form is the exact point of the formulation outlined in your paper, can you comment though ?
