### 2025-05-11 #7831 @knepley
I will respond to the four bullets in order:

1. Yes, I agree that we should use a different symbol for the strong form in (1), maybe $\mathbb{F}$. This is not really a division between strong and weak, but rather a decomposition into parts that depend on the physics (f and F) and things that do not (everything else in $\mathcal{F}$. An identical decomposition can be found [here](https://joss.theoj.org/papers/10.21105/joss.02945), and to my knowledge this decomposition was first developed [here](https://arxiv.org/abs/1309.1204).
2. $B$ is synthesis operation, where a function is projected onto the finite element basis. That matrix is just the evaluation of each basis function for the cell on the quadrature points. The notation $N$ is a crazy thing from mechanical which we are not using. Similarly, D is the evaluation of the basis function gradients at the quadrature points.
3. One is the residual operation and one is the Jacobian. I do not see the necessity of writing them together.
4. Writing (3) allow us to see a mathematical form for the Jacobian, which can be reasoned about and manipulated. SymPy is a particular software strategy for implementing (3). These are categorically different things.
5. FEncCSx and Firedrake use a different decomposition of the weak form. This is why libCEED could not be a backend for either of them.
