### 2025-05-12 #7831 @lmoresi
Updates in b0b9e88fdfd94885d9af157661d0d73192271d15:

  - Address broken references (regression !)
  - Add subscripts to functionals to separate weak and strong forms as pointed out by @chennachaos
  - Whatever the shortcomings of the notation in equations (2) and (3), they reference the original paper by @knepley et al and (we think) readers are best served looking at that document for details of these choices.
  - (2) and (3) can be adjacent if that is easier to follow. The paper has been modified accordingly.
  - We write out the Jacobian term since this is also part of the PETSc template and `sympy` computes the requested terms. In the code there is an almost exact correspondence between the form we write mathematically in (3) and the python code.

@chennachaos asks for clarification of the Fenics / Firedrake codes. We think this is appropriately handled in the *statement of the field* section (missing in the original document) which compares both Fenics / Firedrake as well as other codes with scripted front ends that automatically generate weak forms.

Regarding the *statement of need* section, @chennachaos, are you requesting more domain-specific need (ie. what is missing in geodynamics that underworld3 fixes) or are the changes requested by @gassmoeller sufficient ?
