### 2025-05-13 #7831 @chennachaos
@lmoresi, The documentation looks good and easy to follow. I suggest the following to improve it further.

1. Add the governing equations, key details of formulations and the relevant time schemes used for the notebooks 5, 6, 7 and 8. Cite relevant papers and textbooks as references for the formulations used in the examples.
2. Parallelisation is briefly discussed in the Parallel Execution section. Add parallel scalability in terms of strong and weak scaling for two examples: a scalar field problem (Poisson equation) and a vector field problem (Navier-Stokes). Also, some instructions on changing the PETSc solver options would be helpful.
3. The section on Mesh Adaptation needs expanding a bit. The current image does not clearly depict the adapted mesh. Showing the original mesh and the adapted mesh would be helpful. I understand that it is low-level at present.
4. *https://www.underworlcode.org/* is broken.
