### 2025-05-19 #7831 @lmoresi
@TY-00 — can you look at the comments above as you go through the Quickstart guide and then we'll chat about how to make this easier for a day-one beginner to look at.
