### 2025-05-22 fb11d67
update GitHub action to fail if failures detected previously
