### 2025-05-26 cd0ea2b
Pykdtree (#316)
New pykdtree implementation! Using https://github.com/storpipfugl/pykdtree

Old implementation removed. New one replaces with the same functionality.
i.e. uw.kdtree.KDTree()

Now the class uw.kdtree.KDTree inherits from pykdtree class. Simple cause it's all written in python.

See tests

tests/test_0101_kdtree.py
tests/test_0505_rbf_swarm_mesh.py
for implementation examples.

Now rbf_interpolate() takes p argument. This is the 'power' parameter for the distance weight calculation, i.e. p=1 for 1/distance weighting, p=2 for 1/distance**2 weighting.
e.g. rbf_vel = vel.rbf_interpo
[…]
