+++ environment.yml
@@ -11,6 +11,7 @@ dependencies:
   - mpmath<=1.3
   - mesalib
   - pyvista
+  - pykdtree
   - pytest
   - numpy
   - scipy

[…]
