### 2025-05-26 e751352
Adding pykdtree into setup.cfg
