+++ setup.cfg
@@ -11,6 +11,7 @@ install_requires =
   xxhash
   numpy
   scipy
+  pykdtree
   cython
   sympy >=1.12
   typeguard
