### 2025-05-26 40dc8b7
Disabling pointwise test
