### 2025-05-26 a0a0419
Update publish.yml
Quarto deployment - update version
