+++ .github/workflows/publish.yml
@@ -34,14 +34,14 @@ jobs:
           cd docs/user
           quarto render --cache-refresh --log-level info
 
-      ## Always deploy to version directory
       - name: Deploy quarto book
-        uses: peaceiris/actions-gh-pages@v3
+        uses: peaceiris/actions-gh-pages@v4
 
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           destination_dir: ${{ github.ref_name }}/_quickstart
-          publish_dir: docs/user/_build/book
+          publish_dir: _build
+          keep_files: true
 
       # - name: Deploy quarto lectures
       #   uses: peaceiris/actions-gh-pages@v3
