### 2025-05-26 28f8974
Update publish.yml
Fix publish.yml - update micromamba gh workflow
