### 2025-05-26 9355897
Update publish.yml (#317)
* Does this fix the notebook?
