+++ .github/workflows/publish.yml
@@ -19,15 +19,13 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Add micromamba
-        uses: mamba-org/setup-micromamba@v2
+        uses: mamba-org/setup-micromamba@v1
         with:
-          #micromamba-version: "1.5.6-0" # any version from https://github.com/mamba-org/micromamba-releases
-          #environment-file: .github/workflows/envs/build_quickstart.yml
-          environment: environment.yml
+          micromamba-version: "1.5.6-0" # any version from https://github.com/mamba-org/micromamba-releases
+          environment-file: .github/workflows/envs/build_quickstart.yml
           init-shell: bash
           cache-environment: true
           post-cleanup: "all"
-          create-args: quarto # add quarto
 
       - name: Render docs with quarto
         shell: bash -l {0}
@@ -35,15 +33,14 @@ jobs:
           quarto install tinytex
           cd docs/user
           quarto render --cache-refresh --log-level info
-
+      ## Always deploy to version directory
       - name: Deploy quarto book
-        uses: peaceiris/actions-gh-pages@v4
+        uses: peaceiris/actions-gh-pages@v3
 
         with:
           github_token: ${{ secrets.GITHUB_TOKEN }}
           destination_dir: ${{ github.ref_name }}/_quickstart
-          publish_dir: _build
-          keep_files: true
+          publish_dir: docs/user/_build/book
 
       # - name: Deploy quarto lectures
       #   uses: peaceiris/actions-gh-pages@v3
