### 2025-06-11 2b30360
Changes in the parameter names in the solvers and some other aesthetic changes (#320)
Solver updates to follow Richard - 2015 and documentation tweaks.

---------

Co-authored-by: Louis Moresi <louis.moresi@anu.edu.au>
Co-authored-by: Julian Giordani <julesghub@users.noreply.github.com>
