+++ src/underworld3/systems/solvers.py
@@ -41,7 +41,7 @@ class SNES_Poisson(SNES_Scalar):
     $$
     \nabla \cdot
             \color{Blue}{\underbrace{\Bigl[ \boldsymbol\kappa \nabla u \Bigr]}_{\mathbf{F}}} =
-            \color{Maroon}{\underbrace{\Bigl[ f \Bigl] }_{\mathbf{f}}}
+            \color{Maroon}{\underbrace{\Bigl[ f \Bigl] }_{\mathbf{h}}}
     $$
 
     The term $\mathbf{F}$ relates the flux to gradients in the unknown $u$
@@ -144,10 +144,10 @@ class SNES_Darcy(SNES_Scalar):
     of the Groundwater flow equations
 
     $$
-    \color{Green}{\underbrace{ \Bigl[  S_s \frac{\partial h}{\partial t} \Bigr]}_{\dot{\mathbf{f}}}} -
+    \color{Green}{\underbrace{ \Bigl[  S_s \frac{\partial h}{\partial t} \Bigr]}_{\dot{\mathbf{u}}}} -
     \nabla \cdot
             \color{Blue}{\underbrace{\Bigl[ \boldsymbol\kappa \nabla h  - \boldsymbol{s}\Bigr]}_{\mathbf{F}}} =
-            \color{Maroon}{\underbrace{\Bigl[ W \Bigl] }_{\mathbf{f}}}
+            \color{Maroon}{\underbrace{\Bigl[ W \Bigl] }_{\mathbf{h}}}
     $$
 
     The flux term, $\mathbf{F}$ relates the effective velocity to pressure gradients
@@ -350,7 +350,7 @@ class SNES_Stokes(SNES_Stokes_SaddlePt):
     (or near-incompressibility) constraint.
 
     $$
-    \nabla \cdot
+    -\nabla \cdot
             \color{Blue}{\underbrace{\Bigl[
                     \boldsymbol{\tau} -  p \mathbf{I} \Bigr]}_{\mathbf{F}}} =
             \color{Maroon}{\underbrace{\Bigl[ \mathbf{f} \Bigl] }_{\mathbf{h}}}
@@ -390,7 +390,7 @@ class SNES_Stokes(SNES_Stokes_SaddlePt):
     the term in the `sympy` expression.
 
       - A preconditioner is usually required for the saddle point system and this is provided
-    though the `saddle_preconditioner` property. The default choice is $1/\eta$ for a scalar viscosity function.
+    through the `saddle_preconditioner` property. The default choice is $1/\eta$ for a scalar viscosity function.
 
     ## Notes
 
@@ -622,14 +622,14 @@ class SNES_VE_Stokes(SNES_Stokes):
     All other functionality is inherited from SNES_Stokes
 
     $$
-    \nabla \cdot
+    -\nabla \cdot
             \color{Blue}{\underbrace{\Bigl[
                     \boldsymbol{\tau} -  p \mathbf{I} \Bigr]}_{\mathbf{F}}} =
-            \color{Maroon}{\underbrace{\Bigl[ \mathbf{f} \Bigl] }_{\mathbf{f}}}
+            \color{Maroon}{\underbrace{\Bigl[ \mathbf{f} \Bigl] }_{\mathbf{h}}}
     $$
 
     $$
-    \underbrace{\Bigl[ \nabla \cdot \mathbf{u} \Bigr]}_{\mathbf{f}_p} = 0
+    \underbrace{\Bigl[ \nabla \cdot \mathbf{u} \Bigr]}_{\mathbf{h}_p} = 0
     $$
 
[…]
