### 2025-06-19 d5fc10c
Fix bug in saving mesh HDF5 file (#315)
no 'save' method on mesh

Co-authored-by: Ben Mather <brmather1@gmail.com>
