### 2025-06-24 427dc8d
Update mesh view function with levels (#322)
* Update mesh view function with levels

* Update mesh view with docstring for level option

* Update mesh view docstring
