### 2025-06-24 49c09b8
Alternative to .vtk for mesh to pyvista
The 3.23 petsc viewer for vtk is not recognised by pyvista. This is an alternative strategy that generates the unstructured mesh directly from dmplex. It is less general than the PETSc version in that it assumes the same element type everywhere (which is always true in uw3)
