+++ src/underworld3/visualisation.py
@@ -4,7 +4,6 @@ import os
 
 def initialise(jupyter_backend):
 
-
     import pyvista as pv
 
     pv.global_theme.background = "white"
@@ -27,7 +26,7 @@ def initialise(jupyter_backend):
     return
 
 
-def mesh_to_pv_mesh(mesh, jupyter_backend=None):
+def mesh_to_pv_mesh(mesh0, jupyter_backend=None):
     """Initialise pyvista engine from existing mesh"""
 
     # # Required in notebooks
@@ -40,18 +39,57 @@ def mesh_to_pv_mesh(mesh, jupyter_backend=None):
     import shutil
     import tempfile
     import pyvista as pv
+    import numpy as np
 
-    with tempfile.TemporaryDirectory() as tmp:
-        if type(mesh) == str:  # reading msh file directly
-            vtk_filename = os.path.join(tmp, "tmpMsh.msh")
-            shutil.copyfile(mesh, vtk_filename)
-        else:  # reading mesh by creating vtk
-            vtk_filename = os.path.join(tmp, "tmpMsh.vtk")
-            mesh.vtk(vtk_filename)
+    # with tempfile.TemporaryDirectory() as tmp:
+    #     if type(mesh) == str:  # reading msh file directly
+    #         vtk_filename = os.path.join(tmp, "tmpMsh.msh")
+    #         shutil.copyfile(mesh, vtk_filename)
+    #     else:  # reading mesh by creating vtk
+    #         vtk_filename = os.path.join(tmp, "tmpMsh.vtk")
+    #         mesh.vtk(vtk_filename)
 
-        pvmesh = pv.read(vtk_filename)
+    #     pvmesh = pv.read(vtk_filename)
 
-    return pvmesh
+    # return pvmesh
+
+    ## Alternative - not via file / create an unstructured grid in pyvista
+
+    from petsc4py import PETSc
+
+    match (mesh0.dm.isSimplex(), mesh0.dim):
+        case (True, 2):
+            vtk_cell_type = pv.cell.CellType.TRIANGLE
+        case (True, 3):
+            vtk_cell_type = pv.cell.CellType.TETRA
+        case (False, 2):
+            vtk_cell_type = pv.cell.CellType.QUAD
+        case (False, 3):
+            vtk_cell_type = pv.cell.CellType.HEXAHEDRON
+
+    cStart, cEnd = mesh0.dm.getHeightStratum(0)
+    fStart, fEnd = mesh0.dm.getHeightStratum(1)
+    pStart, pEnd = mesh0.dm.getDepthStratum(0)
+
+    cell_num_points = mesh0.element.entities[mesh0.dim]
+    face_num_points = mesh0.element.face_entities[mesh0.dim]
+
+    cell_points_list = []
+    for cell_id in range(cStart, cEnd):
+        cell_points = mesh0.dm.getTransitiveClosure(cell_id)[0][-cell_num_points:]
+        cell_points_list.append(cell_points - pStart)
+
+    cells_array = np.array(cell_points_list, dtype=int)
+    cells_size = np.full((cells_array.shape[0], 1), cell_num_points, dtype=int)
+    cells_type = np.full((cells_array.shape[0], 1), vtk_cell_type, dtype=int)
+
+    cells_array = np.hstack((cells_size, cells_array), dtype=int)
+
+    pv_mesh = pv.UnstructuredGrid(
+        cells_array, cells_type, coords_to_pv_coords(mesh0.data)
+    )
+
+    return pv_mesh
 
 
 def coords_to_pv_coords(coords):
@@ -243,7 +281,6 @@ def plot_mesh(
     save_png=False,
     dir_fname="",
 ):
-
     """
[…]
