### 2025-06-24 91bec90
"Fix" for pointwise function test
We changed the output in debugging / verbose mode and so one test was failing. It's not a great test because it just checks the output never changes. This will currently work but it is sensitive to the number of solvers instantiated during the test - should replace this with regular expression checking.
