### 2025-06-24 bd8bbb3
Formatting changes
Mostly a results of insisting that the `black` formatter gets run across the source before committing.
