+++ src/underworld3/systems/solvers.py
@@ -19,20 +19,6 @@ from .ddt import SemiLagrangian as SemiLagrangian_DDt
 from .ddt import Lagrangian as Lagrangian_DDt
 
 
-# class UW_Scalar_Temple(SNES_Scalar):
-
-# class UW_Lagrangian_Helper:
-#     """Mixin style ... add some functions to manage swarm updates etc"""
-
-#     @property
-#     def phi_star(self):
-#         return "phi_star"
-
-#     @property
-#     def phi_star_star(self):
-#         return "phi_star_star"
-
-
 class SNES_Poisson(SNES_Scalar):
     r"""
     This class provides functionality for a discrete representation
