+++ src/underworld3/discretisation.py
@@ -2,6 +2,7 @@ from typing import Optional, Tuple, Union
 from enum import Enum
 
 import os
+from mpi4py.MPI import Info
 import numpy
 import sympy
 from sympy.matrices.expressions.blockmatrix import bc_dist
@@ -11,8 +12,10 @@ import underworld3 as uw
 
 from underworld3.utilities._api_tools import Stateful
 from underworld3.utilities._api_tools import uw_object
+from underworld3.utilities._utils import gather_data
 
 from underworld3.coordinates import CoordinateSystem, CoordinateSystemType
+
 from underworld3.cython import petsc_discretisation
 import underworld3.timing as timing
 
@@ -165,6 +168,8 @@ class Mesh(Stateful, uw_object):
         filename=None,
         refinement=None,
         refinement_callback=None,
+        coarsening=None,
+        coarsening_callback=None,
         return_coords_to_bounds=None,
         boundaries=None,
         boundary_normals=None,
@@ -179,11 +184,22 @@ class Mesh(Stateful, uw_object):
         comm = PETSc.COMM_WORLD
 
         if isinstance(plex_or_meshfile, PETSc.DMPlex):
+            isDistributed = plex_or_meshfile.isDistributed()
             if verbose and uw.mpi.rank == 0:
-                print(f"Constructing UW mesh from DMPlex object", flush=True)
+                print(
+                    f"Constructing UW mesh from DMPlex object (distributed == {isDistributed})",
+                    flush=True,
+                )
+            if verbose:
+                plex_or_meshfile.view()
+
             name = "plexmesh"
             self.dm = plex_or_meshfile
             self.sf0 = None  # Should we build one ?
+
+            # Don't set from options — don't want to redistribute the dm
+            # or change any settings as this should be left to the user
+
         else:
             comm = kwargs.get("comm", PETSc.COMM_WORLD)
             name = plex_or_meshfile
@@ -242,6 +258,11 @@ class Mesh(Stateful, uw_object):
 
                 f.close()
 
+                # This needs to be done when reading a dm from a checkpoint
+                # or building from an imported mesh format
+
+                self.dm.setFromOptions()
+
             else:
                 raise RuntimeError(
                     "Mesh file %s has unknown format '%s'."
@@ -276,7 +297,10 @@ class Mesh(Stateful, uw_object):
         # options.delValue("dm_plex_gmsh_mark_vertices")
         # options.delValue("dm_plex_gmsh_multiple_tags")
         # options.delValue("dm_plex_gmsh_use_regions")
-        self.dm.setFromOptions()
+        #
+
+        # Only for newly created dm (from mesh files)
+        # self.dm.setFromOptions()
 
         # uw.adaptivity._dm_stack_bcs(self.dm, self.boundaries, "UW_Boundaries")
 
@@ -315,8 +339,11 @@ class Mesh(Stateful, uw_object):
             uw.mpi.barrier()
 
         ## ---
+        ## Note - coarsening callback is tricky because the coarse meshes do not have the labels
+        ##
 
         self.refinement_callback = refinement_callback
+        self.coarsening_callback = coarsening_callback
         self.name = name
         self.sf1 = None
         self.return_coords_to_bounds = return_coords_to_bounds
@@ -326,6 +353,10 @@ class Mesh(Stateful, uw_object):
         if verbose and uw.mpi.rank == 0:
             print(
                 f"Mesh refinement levels: {refinement}",
+                flush=True,
+            )
+            print(
+                f"Mesh coarsening levels: {coarsening}",
                 flush=True,
             )
 
@@ -359,12 +390,38 @@ class Mesh(Stateful, uw_object):
             self.dm_h = self.dm_hierarchy[-1]
[…]
