### 2025-06-24 4670fe8
Formatting changes
`black` formatter for this pyx file too
