+++ src/underworld3/cython/petsc_discretisation.pyx
@@ -1,6 +1,6 @@
 from typing import Optional, Tuple, Union
 from petsc4py import PETSc
-import underworld3 as uw 
+import underworld3 as uw
 
 # from cython cimport view  # comment this line to see what will happen
 import numpy as np
@@ -15,10 +15,10 @@ include "petsc_extras.pxi"
 def petsc_fvm_get_min_radius(mesh) -> float:
         """
         This method returns the minimum distance from any cell centroid to a face.
-        It wraps to the PETSc `DMPlexGetMinRadius` routine. 
+        It wraps to the PETSc `DMPlexGetMinRadius` routine.
         """
 
-        ## Note: The petsc4py version of DMPlexComputeGeometryFVM does not compute all cells and 
+        ## Note: The petsc4py version of DMPlexComputeGeometryFVM does not compute all cells and
         ## does not obtain the minimum radius for the mesh.
 
         cdef Vec cellgeom = Vec()
@@ -26,7 +26,7 @@ def petsc_fvm_get_min_radius(mesh) -> float:
         cdef DM dm = mesh.dm
 
         DMPlexComputeGeometryFVM(dm.dm,&cellgeom.vec,&facegeom.vec)
-        
+
         min_radius = dm.getMinRadius()
         cellgeom.destroy()
         facegeom.destroy()
@@ -36,10 +36,10 @@ def petsc_fvm_get_min_radius(mesh) -> float:
 def petsc_fvm_get_local_cell_sizes(mesh) -> np.array:
         """
         This method returns the minimum distance from any cell centroid to a face.
-        It wraps to the PETSc `DMPlexGetMinRadius` routine. 
+        It wraps to the PETSc `DMPlexGetMinRadius` routine.
         """
 
-        ## Note: The petsc4py version of DMPlexComputeGeometryFVM does not compute all cells and 
+        ## Note: The petsc4py version of DMPlexComputeGeometryFVM does not compute all cells and
         ## does not obtain the minimum radius for the mesh.
 
         cdef Vec cellgeom = Vec()
@@ -73,20 +73,20 @@ def petsc_dm_create_submesh_from_label(incoming_dm, boundary_label_name, boundar
         cdef PetscBool markedFaces = marked_faces
 
         subdm = PETSc.DM()
-        
+
         DMGetLabel(c_dm.dm, "Boundary", &dmlabel)
         # DMPlexCreateSubmesh(dm.dm, dmlabel, value, markedFaces, &subdm.dm)
 
-        return 
+        return
 
-        
 
-# This is not cython, does it need to be here or in discretisation.py ?       
+
+# This is not cython, does it need to be here or in discretisation.py ?
 
 def petsc_dm_find_labeled_points_local(dm, label_name, sectionIndex=False, verbose=False):
-        '''Identify local points associated with "Label" 
-        
-        dm -> expects a petscDM object 
+        '''Identify local points associated with "Label"
+
+        dm -> expects a petscDM object
         label_name -> "String Name for Label"
         sectionIndex -> False: leave points as indexed by the relevant section on the dm
                         True: index into the local coordinate array
@@ -103,7 +103,7 @@ def petsc_dm_find_labeled_points_local(dm, label_name, sectionIndex=False, verbo
                 if uw.mpi.rank == 0:
                         print(f"Label {label_name} is not present on the dm")
                 return np.array([0])
-                
+
         pointIS = dm.getStratumIS("depth",0)
[…]
