### 2025-06-24 4841242
BASIC swarm
Adding a "BASIC_Swarm" type. Currently a clone of the standard (PIC) swarm but without any cell_id references. I keep the same (PETSc) field name for the coordinates as in the PIC swarm for compatibility.

The swarm now has a `migrate` function which handles the communication using kdtree based redistribution and kdtree based checking of processor ownership

Currently missing: an equivalent of the cell-by-cell `populate` method which would still be useful even in the absence of PIC functionality
