+++ src/underworld3/swarm.py
@@ -19,10 +19,13 @@ comm = uw.mpi.comm
 from enum import Enum
 
 
+# We can grab this type from the PETSc module
 class SwarmType(Enum):
+    DMSWARM_BASIC = 0
     DMSWARM_PIC = 1
 
 
+# We can grab this type from the PETSc module
 class SwarmPICLayout(Enum):
     """
     Particle population fill type:
@@ -518,11 +521,14 @@ class SwarmVariable(Stateful, uw_object):
 
         # check if swarmFilename exists
         if os.path.isfile(os.path.abspath(swarmFilename)):  # easier to debug abs path
+            print(f"Reading swarm information from {swarmFilename}", flush=True)
             pass
         else:
             raise RuntimeError(f"{os.path.abspath(swarmFilename)} does not exist")
 
         if os.path.isfile(os.path.abspath(filename)):
+            print(f"Reading variable information from {filename}", flush=True)
+
             pass
         else:
             raise RuntimeError(f"{os.path.abspath(filename)} does not exist")
@@ -549,12 +555,12 @@ class SwarmVariable(Stateful, uw_object):
                 all_coords = h5f_swarm["coordinates"][()]
                 all_data = h5f_data["data"][()]
 
-                cell = self.swarm.mesh.get_closest_local_cells(all_coords)
-                local = np.where(cell >= 0)[0]
-                # not_not_local = np.where(cell == -1)[0]
+                # cell = self.swarm.mesh.get_closest_local_cells(all_coords)
+                # local = np.where(cell >= 0)[0]
+                # # not_not_local = np.where(cell == -1)[0]
 
-                local_coords = all_coords[local]
-                local_data = all_data[local]
+                local_coords = all_coords  # [local]
+                local_data = all_data  # [local]
 
                 kdt = uw.kdtree.KDTree(local_coords)
 
@@ -591,7 +597,7 @@ class IndexSwarmVariable(SwarmVariable):
         self.radius_s = radius**2
         self.update_type = update_type
         if self.update_type == 1:
-            self.nnn_bc = npoints_bc 
+            self.nnn_bc = npoints_bc
             self.ind_bc = ind_bc
 
         # These are the things we require of the generic swarm variable type
@@ -702,7 +708,7 @@ class IndexSwarmVariable(SwarmVariable):
             2) for each index in the set, we create a mask mesh variable by mapping 1.0 wherever the
                index matches and 0.0 where it does not.
 
-        NOTE: If no material is identified with a given nodal value, the default is to impose 
+        NOTE: If no material is identified with a given nodal value, the default is to impose
         a near-neighbour hunt for a valid material and set that one
 
         ## ToDo: This should be revisited to match the updated master copy of _update
@@ -714,46 +720,52 @@ class IndexSwarmVariable(SwarmVariable):
         if self.update_type == 0:
             kd = uw.kdtree.KDTree(self._meshLevelSetVars[0].coords)
             kd.build_index()
-            
+
             with self.swarm.access():
-                n_indices, n_distance = kd.find_closest_n_points(self.nnn,self.swarm.particle_coordinates.data)
+                n_indices, n_distance = kd.find_closest_n_points(
+                    self.nnn, self.swarm.particle_coordinates.data
+                )
                 kd_swarm = uw.kdtree.KDTree(self.swarm.particle_coordinates.data)
                 kd_swarm.build_index()
                 n, d, b = kd_swarm.find_closest_point(self._meshLevelSetVars[0].coords)
-        
+
             for ii in range(self.indices):
                 meshVar = self._meshLevelSetVars[ii]
-            
+
[…]
