### 2025-06-24 a53470b
Hiding some functions
