+++ src/underworld3/function/__init__.py
@@ -5,6 +5,8 @@ from ._function import (
     evalf,
     dm_swarm_get_migrate_type,
     dm_swarm_set_migrate_type,
+    _dmswarm_get_migrate_type,
+    _dmswarm_set_migrate_type,
     # evalf_at_coords,
     # _interpolate_all_vars_on_mesh,
 )
