### 2025-06-25 3f02797
Diffusion bits (#311)
* Diffusion updates

solver updates
-Add a diffusion solver
-symbolic ddt
-Eulerian ddt
CM updates
-Anisotropic diffusion CM
-Flux CM, allowing flux terms to be passed directly into the solver

* Change to symbolic notation

Functions now appear correctly in the view() function

* Add L2_norm calculation

* Remove separate diffusion_solver file

Folded into solvers.py

* Update advDiff test

Aligns with the already updated test in the joss paper branch

---------

Co-authored-by: Julian Giordani <julesghub@users.noreply.github.com>
