+++ tests/test_1100_AdvDiffCartesian.py
@@ -6,31 +6,32 @@
 
 import underworld3 as uw
 import numpy as np
+import sympy as sp
 import math
 import pytest
+import matplotlib.pyplot as plt
 
 
 # ### Set up variables of the model
 
 # +
-res = 12
+res = 24
 nsteps = 1
 kappa = 1.0  # diffusive constant
 
 velocity = 1 / res  # /res
 
-### min and max temps
-tmin = 0.5  # temp min
-tmax = 1.0  # temp max
 
-### Thickness of hot pipe in centre of box
-pipe_thickness = 0.4
-
+t_start = 1e-4
+t_end   = 2e-4
 
 # ### Set up the mesh
-
 xmin, xmax = 0, 1
 ymin, ymax = 0, 1
+
+u_degree = 2
+
+
 
 ### Quads
 meshStructuredQuadBox = uw.meshing.StructuredQuadBox(
@@ -47,8 +48,20 @@ unstructured_simplex_box_irregular = uw.meshing.UnstructuredSimplexBox(
 unstructured_simplex_box_regular = uw.meshing.UnstructuredSimplexBox(
     cellSize=1 / res, regular=True, qdegree=3, refinement=0
 )
+# -
+
+# ### setup analytical function
+
+# +
+u, t, x, x0, x1 = sp.symbols('u, t, x, x0, x1')
 
 
+U_a_x = 0.5 * ( sp.erf( (x1  - x+(u*t))  / (2*sp.sqrt(kappa*t))) + sp.erf( (-x0 + x-(u*t))  / (2*sp.sqrt(kappa*t))) )
+
+
+
+
+# +
 @pytest.mark.parametrize(
     "mesh",
     [
@@ -62,7 +75,7 @@ def test_advDiff_boxmesh(mesh):
 
     # Create an mesh vars
     v = uw.discretisation.MeshVariable("U", mesh, mesh.dim, degree=1)
-    T = uw.discretisation.MeshVariable("T", mesh, 1, degree=3)
+    T = uw.discretisation.MeshVariable("T", mesh, 1, degree=u_degree)
 
     # #### Create the advDiff solver
 
@@ -82,102 +95,48 @@ def test_advDiff_boxmesh(mesh):
     adv_diff.constitutive_model.Parameters.diffusivity = kappa
 
     ### fix temp of top and bottom walls
-    adv_diff.add_dirichlet_bc(tmin, "Bottom")
-    adv_diff.add_dirichlet_bc(tmin, "Top")
-    # adv_diff.add_dirichlet_bc(0., "Left")
-    # adv_diff.add_dirichlet_bc(0., "Right")
+    adv_diff.add_dirichlet_bc(0., "Left")
+    adv_diff.add_dirichlet_bc(0., "Right")
 
     with mesh.access(v):
         # initialise fields
         # v.data[:,0] = -1*v.coords[:,1]
         v.data[:, 1] = velocity
 
+    U_start = U_a_x.subs({u:velocity, t:t_start,x:mesh.X[0], x0:0.4, x1:0.6})
     with mesh.access(T):
-        T.data[...] = tmin
+        T.data[:,0] = uw.function.evaluate(U_start, T.coords)
 
-        pipePosition = ((ymax - ymin) - pipe_thickness) / 2.0
 
-        T.data[
-            (T.coords[:, 1] >= (T.coords[:, 1].min() + pipePosition))
-            & (T.coords[:, 1] <= (T.coords[:, 1].max() - pipePosition))
-        ] = tmax
-
-    # ### Create points to sample the UW results
-    ### y coords to sample
-    sample_y = np.arange(
[…]
