+++ src/underworld3/maths/functions.py
@@ -4,6 +4,7 @@ import sympy
 from sympy import sympify
 from typing import Optional, Callable
 from underworld3 import function
+from underworld3 import maths
 
 
 def delta(
@@ -15,6 +16,34 @@ def delta(
 
     return delta_fn
 
+def L2_norm(n_s, a_s, mesh):
+    """
+    Compute the L2 norm (Euclidean norm) of the difference between numerical and analytical solutions.
+    
+    Parameters:
+    n_s   : Numeric solution (scalar or vector field)
+    a_s   : Analytic solution (scalar or vector field)
+    mesh  : The mesh used for integration
+
+    Returns:
+    L2 norm value (for scalar or vector)
+    """
+    # Check if the input is a vector (SymPy Matrix)
+    if isinstance(n_s, sympy.Matrix) and n_s.shape[1] > 1:
+        # Compute squared difference using dot product for vectors
+        squared_difference = (n_s - a_s).dot(n_s - a_s)
+    else:
+        # Compute squared difference for scalars
+        squared_difference = (n_s - a_s) ** 2
+
+
+    # Integral over the domain
+    I = maths.Integral(mesh, squared_difference)
+
+    # Compute the L2 norm
+    L2 = sympy.sqrt( I.evaluate() )
+
+    return L2
 
 # Definitions of expressions that are used in various places within underworld
 
