### 2025-06-25 c0a588e
updated checkpoint_xdmf to look for correct vertex field name in h5 file (#324)
* updated checkpoint_xdmf to look for correct vertex field name in h5 file

* added test to check vertex fields in xdmf and h5 files

* minor updates to the test
