### 2025-06-26 618f039
Get all local points from DM label (#319)
* get all local points from DM label

* get rid of if statements - empty indices break the for loop anyway

* revert to numpy intersect1d + union1d and make parallel-safe

* consider using DMPlexDistributeOverlap in future for this "puff" algorithm
---------

Co-authored-by: Ben Mather <brmather1@gmail.com>
