### 2025-06-27 aa65291
added var as argument to def
