### 2025-06-28 8d6c94b
Advection for BASIC swarm
Note: in PETSc 3.23.x, the PIC swarm interface has changed and it is no longer possible to access the CellID from petsc4py. Failing tests are now expected in standard swarms and nodal-point swarms. The goal is to make the BASIC swarm the default and use PIC only when necessary (never ?)
