+++ src/underworld3/swarm.py
@@ -2284,70 +2284,67 @@ class BASIC_Swarm(Stateful, uw_object):
         if global_unclaimed_points == 0:
             return
 
-        for it in range(0, min(max_its, uw.mpi.size)):
+        # Migrate particles between processors if appropriate
+        # Otherwise skip the next step and just remove missing points
+        # and tidy up.
 
-            # Send unclaimed points to next processor in line
+        if uw.mpi.size > 1:
+            for it in range(0, min(max_its, uw.mpi.size)):
 
-            swarm_rank_array = self.dm.getField("DMSwarm_rank")
-            swarm_coord_array = self.dm.getField("DMSwarmPIC_coor").reshape(
-                (-1, self.dim)
-            )
+                # Send unclaimed points to next processor in line
 
-            # Should be able to do this only for unclaimed points
-            if len(swarm_coord_array > 0):
-                rank, dist = mesh_domain_kdtree.find_closest_n_points(
-                    it + 1, swarm_coord_array[not_my_points]
+                swarm_rank_array = self.dm.getField("DMSwarm_rank")
+                swarm_coord_array = self.dm.getField("DMSwarmPIC_coor").reshape(
+                    (-1, self.dim)
                 )
-                swarm_rank_array[not_my_points, 0] = rank[:, it]
 
-            self.dm.restoreField("DMSwarmPIC_coor")
-            self.dm.restoreField("DMSwarm_rank")
+                # Should be able to do this only for unclaimed points
+                if len(swarm_coord_array > 0):
+                    rank, dist = mesh_domain_kdtree.find_closest_n_points(
+                        it + 1, swarm_coord_array[not_my_points]
+                    )
+                    swarm_rank_array[not_my_points, 0] = rank[:, it]
 
-            # Now we send the points (basic migration)
-            self.dm.migrate(remove_sent_points=True)
-            uw.mpi.barrier()
+                self.dm.restoreField("DMSwarmPIC_coor")
+                self.dm.restoreField("DMSwarm_rank")
 
-            swarm_coord_array = self.dm.getField("DMSwarmPIC_coor").reshape(
-                (-1, self.dim)
-            )
+                # Now we send the points (basic migration)
+                self.dm.migrate(remove_sent_points=True)
+                uw.mpi.barrier()
 
-            in_or_not = self.mesh.points_in_domain(swarm_coord_array)
-            self.dm.restoreField("DMSwarmPIC_coor")
+                swarm_coord_array = self.dm.getField("DMSwarmPIC_coor").reshape(
+                    (-1, self.dim)
+                )
 
-            num_points_in_domain = np.count_nonzero(in_or_not == True)
-            num_points_not_in_domain = np.count_nonzero(in_or_not == False)
-            not_my_points = np.where(in_or_not == False)[0]
+                in_or_not = self.mesh.points_in_domain(swarm_coord_array)
+                self.dm.restoreField("DMSwarmPIC_coor")
 
-            unclaimed_points_last_iteration = global_unclaimed_points
-            claimed_points_last_iteration = global_claimed_points
+                num_points_in_domain = np.count_nonzero(in_or_not == True)
+                num_points_not_in_domain = np.count_nonzero(in_or_not == False)
+                not_my_points = np.where(in_or_not == False)[0]
 
-            global_unclaimed_points = int(
-                uw.utilities.gather_data(
-                    num_points_not_in_domain,
-                    bcast=True,
-                    dtype=int,
-                ).sum()
-            )
+                unclaimed_points_last_iteration = global_unclaimed_points
+                claimed_points_last_iteration = global_claimed_points
 
-            global_claimed_points = int(
-                uw.utilities.gather_data(
-                    num_points_in_domain, bcast=True, dtype=int
-                ).sum()
-            )
+                global_unclaimed_points = int(
[…]
