### 2025-06-28 2722460
kdtree updates propagating into swarm rework code
Not sure if I have caught all the interface changes, but the ones in the tests I have found / fixed
