+++ src/underworld3/discretisation.py
@@ -588,7 +588,7 @@ class Mesh(Stateful, uw_object):
         The mesh dimensionality.
         """
         return self.dm.getCoordinateDim()
-      
+
     @property
     def element(self) -> dict:
         """
@@ -599,40 +599,46 @@ class Mesh(Stateful, uw_object):
         return self._element
 
     def view(self, level=0):
-        '''
+        """
         Displays mesh information at different levels.
-        
+
         Parameters
         ----------
-        level : int (0 default) 
-            The display level. 
+        level : int (0 default)
+            The display level.
             0, for basic mesh information (variables and boundaries), while level=1 displays detailed mesh information (including PETSc information)
-        '''
-
+        """
 
         import numpy as np
-        if level==0:
+
+        if level == 0:
             if uw.mpi.rank == 0:
                 print(f"\n")
                 print(f"Mesh # {self.instance}: {self.name}\n")
 
                 uw.visualisation.plot_mesh(self)
 
-                #Total number of cells
-                nstart, nend=self.dm.getHeightStratum(0)
-                num_cells=nend-nstart
+                # Total number of cells
+                nstart, nend = self.dm.getHeightStratum(0)
+                num_cells = nend - nstart
                 print(f"Number of cells: {num_cells}\n")
 
                 if len(self.vars) > 0:
-                    print(f"| Variable Name       | component | degree |     type        |")
-                    print(f"| ---------------------------------------------------------- |")
+                    print(
+                        f"| Variable Name       | component | degree |     type        |"
+                    )
+                    print(
+                        f"| ---------------------------------------------------------- |"
+                    )
                     for vname in self.vars.keys():
                         v = self.vars[vname]
                         print(
                             f"| {v.clean_name:<20}|{v.num_components:^10} |{v.degree:^7} | {v.vtype.name:^15} |"
                         )
 
-                    print(f"| ---------------------------------------------------------- |")
+                    print(
+                        f"| ---------------------------------------------------------- |"
+                    )
                     print("\n", flush=True)
                 else:
                     print(f"No variables are defined on the mesh\n", flush=True)
@@ -705,28 +711,33 @@ class Mesh(Stateful, uw_object):
             # self.dm.view()
             print(f"Use view(1) to view detailed mesh information.\n")
 
-
-        elif level==1:
+        elif level == 1:
             if uw.mpi.rank == 0:
                 print(f"\n")
                 print(f"Mesh # {self.instance}: {self.name}\n")
                 uw.visualisation.plot_mesh(self)
 
-                #Total number of cells
-                nstart, nend=self.dm.getHeightStratum(0)
-                num_cells=nend-nstart
+                # Total number of cells
+                nstart, nend = self.dm.getHeightStratum(0)
+                num_cells = nend - nstart
                 print(f"Number of cells: {num_cells}\n")
 
                 if len(self.vars) > 0:
-                    print(f"| Variable Name       | component | degree |     type        |")
-                    print(f"| ---------------------------------------------------------- |")
+                    print(
+                        f"| Variable Name       | component | degree |     type        |"
+                    )
+                    print(
+                        f"| ---------------------------------------------------------- |"
+                    )
                     for vname in self.vars.keys():
                         v = self.vars[vname]
                         print(
[…]

[…]
