+++ environment.yml
@@ -4,31 +4,33 @@ channels:
 dependencies:
   - python=3.11
   - compilers
-  - openmpi
-  - openmpi-mpicc
-  - openmpi-mpicxx
-  - petsc=3.21.5 # Fix for now (errors with hdf5 / uw3)
+  - mpich
+  - petsc=3.21.5
   - petsc4py=3.21.5
   - cython=3.*
   - mpmath<=1.3
   - mesalib
+  - pyvista
+  - pykdtree
+  - pytest
   - numpy
   - scipy
-  - mpi4py
-  - h5py
   - sympy
-  - pytest
   - ipython
-  - pyvista
+  - mpi4py
+  - h5py=*=mpi*
+  - trame
+  - ipywidgets
+  - gmsh
+  - python-gmsh
+  - python-xxhash
+  - jupyterlab
+  - jupytext
+  - trame-vtk
+  - trame-vuetify
+  - pygments
+  - pint
+  - typeguard
   - pip
-  # gmsh dependencies (prefer not to use apt.txt)
-  - xorg-libxft
-  - xorg-libxinerama
-  - pip:
-      - gmsh
-      - gmsh-api
-      - jupyterlab
-      - jupytext
-      - pygments
-      - trame-vtk
-      - trame-vuetify
+  # - pip:
+  #   - gmsh-api # is this used still in UW?
