### 2025-06-30 faed46c
updated checkpoint_xdmf method save discontinuous fields properly (#1)
* updated checkpoint_xdmf method save discontinuous fields properly

* added var as argument to def
