### 2025-07-01 b1b8dca
Mesh class: clean up code and update KDTree
Removed commented-out and unused code blocks for clarity. Updated KDTree method calls to use the 'query' interface with appropriate arguments, improving consistency and potentially performance.
