### 2025-07-01 4cc1209
Add docstrings to swarm classes. NodalPointBasicSwarm
Added comprehensive docstrings to Swarm and UW_Swarm classes, clarifying their usage, parameters, and methods. Renamed BASIC_Swarm to UW_Swarm and updated its documentation. Introduced NodalPointBasicSwarm as a nodal swarm implementation based on UW_Swarm. Updated kdtree query usage and removed obsolete comments and code.
