+++ src/underworld3/swarm.py
@@ -332,7 +332,6 @@ class SwarmVariable(Stateful, uw_object):
         kd = uw.kdtree.KDTree(meshVar.coords)
 
         with self.swarm.access():
-            # n, d, b = kd.find_closest_point(self.swarm.data)
             d, n = kd.query(self.swarm.data, k=1)
 
             node_values = np.zeros((meshVar.coords.shape[0], self.num_components))
@@ -721,13 +720,14 @@ class IndexSwarmVariable(SwarmVariable):
             kd = uw.kdtree.KDTree(self._meshLevelSetVars[0].coords)
 
             with self.swarm.access():
-                # n_indices, n_distance = kd.find_closest_n_points(self.nnn,self.swarm.particle_coordinates.data)
                 n_distance, n_indices = kd.query(
                     self.swarm.particle_coordinates.data, k=self.nnn
                 )
                 kd_swarm = uw.kdtree.KDTree(self.swarm.particle_coordinates.data)
                 # n, d, b = kd_swarm.find_closest_point(self._meshLevelSetVars[0].coords)
-                d, n = kd_swarm.query(self._meshLevelSetVars[0].coords, k=1)
+                d, n = kd_swarm.query(
+                    self._meshLevelSetVars[0].coords, k=1, sqr_dist=True
+                )
 
             for ii in range(self.indices):
                 meshVar = self._meshLevelSetVars[ii]
@@ -762,9 +762,8 @@ class IndexSwarmVariable(SwarmVariable):
         elif self.update_type == 1:
             with self.swarm.access():
                 kd = uw.kdtree.KDTree(self.swarm.particle_coordinates.data)
-                # n_indices, n_distance = kd.find_closest_n_points(self.nnn,self._meshLevelSetVars[0].coords)
                 n_distance, n_indices = kd.query(
-                    self._meshLevelSetVars[0].coords, k=self.nnn
+                    self._meshLevelSetVars[0].coords, k=self.nnn, sqr_dist=True
                 )
 
             for ii in range(self.indices):
@@ -810,6 +809,106 @@ class IndexSwarmVariable(SwarmVariable):
 
 
 class Swarm(Stateful, uw_object):
+    """
+    Particle swarm implementation with automatic mesh-particle interactions.
+
+    The `Swarm` class is Underworld's primary particle management system, built on PETSc's
+    DMSWARM_PIC type. It provides automatic particle migration, mesh-particle connectivity,
+    and streamlined particle operations for Lagrangian particle tracking and data storage.
+
+    Differences from UW_Swarm:
+    - **Mesh Integration**: Built-in particle-in-cell (PIC) connectivity with automatic cell tracking
+    - **Migration**: Uses the standard PETSc strategy for migration which depends on the DM type. This requires
+      calculation of cell-relationships each time the coordinates are updated and particles that are not found will
+      be deleted.
+
+
+    Parameters
+    ----------
+    mesh : uw.discretisation.Mesh
+        The mesh object that defines the computational domain. Particles will be
+        automatically associated with mesh cells for efficient spatial operations.
+    recycle_rate : int, optional
+        Rate at which particles are recycled for streak management. If > 1, enables
+        streak particle functionality where particles are duplicated and tracked
+        across multiple cycles. Default is 0 (no recycling).
+    verbose : bool, optional
+        Enable verbose output for debugging and monitoring particle operations.
+        Default is False.
+
+    Attributes
+    ----------
+    mesh : uw.discretisation.Mesh
+        Reference to the associated mesh object.
+    dim : int
+        Spatial dimension of the mesh (2D or 3D).
+    cdim : int
+        Coordinate dimension of the mesh.
[…]
