### 2025-07-03 ee394d2
NodalSwarm rework
Introducing a NodalPointUWSwarm which uses the new BASIC swarm in place of the PIC version. Adjusted the ddt semi-Lagrange solver to use this and checking that it works with advection/diffusion.

Updated the uw.function.evaluate to do this:

  - check points that are in / out of the domain
  - use the petsc interpolation for points in the domain
  - use evalf rbf extrapolation for points outside the domain.

Adding an over-ride _evalf argument for the adv-diffusion solve to force the extrapolation / rbf everywhere.

This is a placeholder commit. There are still bits of string
[…]
