+++ src/underworld3/systems/solvers.py
@@ -1442,6 +1442,7 @@ class SNES_AdvectionDiffusion(SNES_Scalar):
         zero_init_guess: bool = True,
         timestep: float = None,
         _force_setup: bool = False,
+        _evalf=False,
         verbose=False,
     ):
         """
@@ -1471,19 +1472,19 @@ class SNES_AdvectionDiffusion(SNES_Scalar):
 
         # Update History / Flux History terms
         # SemiLagrange and Lagrange may have different sequencing.
-        self.DuDt.update_pre_solve(timestep, verbose=verbose)
-        self.DFDt.update_pre_solve(timestep, verbose=verbose)
+        self.DuDt.update_pre_solve(timestep, verbose=verbose, evalf=_evalf)
+        self.DFDt.update_pre_solve(timestep, verbose=verbose, evalf=_evalf)
 
         super().solve(zero_init_guess, _force_setup)
 
-        self.DuDt.update_post_solve(timestep, verbose=verbose)
-        self.DFDt.update_post_solve(timestep, verbose=verbose)
+        self.DuDt.update_post_solve(timestep, verbose=verbose, evalf=_evalf)
+        self.DFDt.update_post_solve(timestep, verbose=verbose, evalf=_evalf)
 
         self.is_setup = True
         self.constitutive_model._solver_is_setup = True
 
         return
-    
+
 
 class SNES_Diffusion(SNES_Scalar):
     r"""
@@ -1496,7 +1497,7 @@ class SNES_Diffusion(SNES_Scalar):
             \color{Maroon}{\underbrace{\Bigl[ f \Bigl] }_{\mathbf{f}}}
     $$
 
-    The term $\mathbf{F}$ relates diffusive fluxes to gradients in the unknown $u$. 
+    The term $\mathbf{F}$ relates diffusive fluxes to gradients in the unknown $u$.
 
     ## Properties
 
@@ -1529,7 +1530,7 @@ class SNES_Diffusion(SNES_Scalar):
         mesh: uw.discretisation.Mesh,
         u_Field: uw.discretisation.MeshVariable,
         order: int = 1,
-        theta: float = 0.,
+        theta: float = 0.0,
         evalf: Optional[bool] = False,
         verbose=False,
         DuDt: Union[Eulerian_DDt, SemiLagrangian_DDt, Lagrangian_DDt] = None,
@@ -1594,12 +1595,11 @@ class SNES_Diffusion(SNES_Scalar):
 
         if DFDt is None:
             self.Unknowns.DFDt = Symbolic_DDt(
-                sympy.Matrix(
-                    [[0] * self.mesh.dim] ),
-                    varsymbol=rf"{{F[ {self.u.symbol} ] }}",
-                    theta=theta,
-                    bcs=None,
-                    order=order,
+                sympy.Matrix([[0] * self.mesh.dim]),
+                varsymbol=rf"{{F[ {self.u.symbol} ] }}",
+                theta=theta,
+                bcs=None,
+                order=order,
             )
             ### solution unable to solve after n timesteps, due to the projection of flux term (???)
             # self.Unknowns.DFDt = Eulerian_DDt(
@@ -1626,7 +1626,7 @@ class SNES_Diffusion(SNES_Scalar):
 
         f0 = uw.function.expression(
             r"f_0 \left( \mathbf{u} \right)",
-            -self.f + sympy.simplify( self.DuDt.bdf() ) / self.delta_t,
+            -self.f + sympy.simplify(self.DuDt.bdf()) / self.delta_t,
             "Diffusion pointwise force term: f_0(u)",
         )
 
@@ -1658,7 +1658,6 @@ class SNES_Diffusion(SNES_Scalar):
         self.is_setup = False
         self._f = sympy.Matrix((value,))
 
-
[…]

[…]
