### 2025-07-07 #7831 @diehlpk
@lmoresi if I do not hear back by you this Friday, I assume you retracted the paper.
