### 2025-07-07 0778b7d
Fix uw.function.evaluate / eliminate evalf
Rewrite of the uw.function.evaluate code. Now the code will check for points inside the domain and use PETSc interpolation for those. Points outside the domain are computed using the rbf_interpolation code which is perfectly happy to extrapolate wildly. This prevents crashes we were seeing with NodalPointSwarms / Semi-Lagrangian computations (particularly in curved domains).

Rather than exposing the evalf as a separate function, we now supply an argument to the evaluate method. We might want to change the name to rbf or something intuitive / meaningf
[…]
