+++ src/underworld3/systems/ddt.py
@@ -327,16 +327,18 @@ class Eulerian(uw_object):
                         self.psi_star[0].data[...] = self._psi_meshVar.data[...]
                     # print('copying data', flush=True)
                 except:
-                    if self.evalf:
-                        self.psi_star[0].data[...] = uw.function.evalf(
-                            self.psi_fn, self.psi_star[0].coords
-                        ).reshape(-1, max(self.psi_fn.shape))
-                        # print('evalf data', flush=True)
-                    else:
-                        self.psi_star[0].data[...] = uw.function.evaluate(
-                            self.psi_fn, self.psi_star[0].coords
-                        ).reshape(-1, max(self.psi_fn.shape))
-                        # print('evaluate data', flush=True)
+                    # if self.evalf:
+                    #     self.psi_star[0].data[...] = uw.function.evalf(
+                    #         self.psi_fn, self.psi_star[0].coords
+                    #     ).reshape(-1, max(self.psi_fn.shape))
+                    #     # print('evalf data', flush=True)
+                    # else:
+                    self.psi_star[0].data[...] = uw.function.evaluate(
+                        self.psi_fn,
+                        self.psi_star[0].coords,
+                        evalf=evalf,
+                    ).reshape(-1, max(self.psi_fn.shape))
+                    # print('evaluate data', flush=True)
 
         except:
             self._setup_projections()
@@ -674,8 +676,8 @@ class SemiLagrangian(uw_object):
                 order=1,
                 corrector=False,
                 restore_points_to_domain_func=self.mesh.return_coords_to_bounds,
-                evalf=False,
-                step_limit=True,
+                evalf=evalf,
+                step_limit=False,
                 #! substepping: this seems to be too diffusive if left on.
                 #! Check the code carefully !
             )
@@ -686,30 +688,31 @@ class SemiLagrangian(uw_object):
                 # is required instead.
 
                 try:
-                    # Note, we should honour the evalf flag here too.
                     with self._workVar.mesh.access():
                         self.psi_star[0].data[...] = uw.function.evaluate(
-                            self.psi_fn, self.psi_star[0].coords
+                            self.psi_fn,
+                            self.psi_star[0].coords,
+                            evalf=evalf,
                         )
                 except:
                     self._psi_star_projection_solver.uw_function = self.psi_fn
                     self._psi_star_projection_solver.smoothing = 0.0
                     self._psi_star_projection_solver.solve(verbose=verbose)
 
-            if evalf:
-                with self._nswarm_psi.access(self._nswarm_psi.swarmVariable):
-                    for d in range(self.psi_star[i].shape[1]):
-                        self._nswarm_psi.swarmVariable.data[:, d] = uw.function.evalf(
-                            self.psi_star[i].sym[d], self._nswarm_psi.data
-                        )
-            else:
-                with self._nswarm_psi.access(self._nswarm_psi.swarmVariable):
-                    for d in range(self.psi_star[i].shape[1]):
-                        self._nswarm_psi.swarmVariable.data[:, d] = (
-                            uw.function.evaluate(
-                                self.psi_star[i].sym[d], self._nswarm_psi.data
-                            )
-                        )
+            # if evalf:
+            #     with self._nswarm_psi.access(self._nswarm_psi.swarmVariable):
+            #         for d in range(self.psi_star[i].shape[1]):
+            #             self._nswarm_psi.swarmVariable.data[:, d] = uw.function.evalf(
[…]
