+++ src/underworld3/function/__init__.py
@@ -2,7 +2,6 @@ from . import analytic
 from ._function import (
     UnderworldFunction,
     evaluate,
-    evalf,
     dm_swarm_get_migrate_type,
     dm_swarm_set_migrate_type,
     _dmswarm_get_migrate_type,
