+++ src/underworld3/function/_function.pyx
@@ -164,13 +164,13 @@ class UnderworldFunction(sympy.Function):
         return ourcls
 
 
-
 def evaluate(   expr,
                 np.ndarray coords=None,
                 coord_sys=None,
                 other_arguments=None,
                 simplify=True,
-                verbose=False, ):
+                verbose=False,
+                evalf=False):
     """
     Evaluate a given expression at a list of coordinates.
 
@@ -190,28 +190,12 @@ def evaluate(   expr,
         Dictionary of other arguments necessary to evaluate function.
         Not yet implemented.
 
-    Notes
-    -----
-    This function leverages Sympy's `lambdify` function to provide efficient
-    expression evaluation. It operates as follows:
-        1. Extract all Underworld variables functions from the expression. Note that
-           all variables functions must be leaf nodes of the corresponding expression
-           tree, as the variable function arguments must simply be the coordinate
-           vector `mesh.r`. This is a necessary requirement to avoid complication in the
-           domain decomposed parallel runtime situation, where a modified variable function
-           argument (such as `mesh.r - (10,0)`) might translate the variable function onto
-           a neighbouring subdomain. Handling this would result in great complication and
-           inefficiency, and we therefore disallow it.
-        2. Each variable function is evaluated at the user provided coordinates to generate
-           an array of evaluated results.
-        3. Replace all variable function instances within the expression with sympy
-           symbol placeholders.
-        4. Generate a Sympy lambdified expression. This expression takes as arguments the
-           user provided coordinates, and the Underworld variable function placeholders.
-        5. Evaluate the generated lambdified expresson using the coordinate array and
-           evaluated variable function result arrays.
-        6. Return results array for full expression evaluation.
+
     """
+
+
+    # Extract all the mesh/swarm variables in the expression and check that they all live on the
+    # same mesh. If not, this evaluation is not valid.
 
     varfns = set()
     def unpack_var_fns(exp):
@@ -231,8 +215,12 @@ def evaluate(   expr,
             for sub_exp in exp:
                 if isinstance(sub_exp, uw.function._function.UnderworldAppliedFunction):
                     varfns.add(sub_exp)
+                else:
+                    for arg in sub_exp.args:
+                        unpack_var_fns(arg)
+
         else:
-            # Recurse.
+            # Recursively search for more functions
             for arg in exp.args:
                 unpack_var_fns(arg)
 
@@ -240,64 +228,84 @@ def evaluate(   expr,
 
     unpack_var_fns(expr)
 
+
+    if verbose:
+        for varfn in varfns:
+            print(f"MeshVariable functions in evaluation: {varfns}")
+
+
     # Check the same mesh is used for all mesh variables
     mesh = None
     for varfn in varfns:
-
         if mesh is None:
             mesh = varfn.meshvar().mesh
-            #mesh = varfn.mesh
         else:
             if mesh != varfn.meshvar().mesh:
                 raise RuntimeError("In this expression there are functions defined on different meshes. This is not supported")
 
-    if mesh is not None:
-        print(f"Mesh for evaluations: {mesh.name}", flush=True)
 
-        in_or_not = mesh.points_in_domain(coords)
+    # If there are no mesh variables, then we have no need of a mesh to
+    # help us to evaluate the expression. The evalf flag will force rbf_evaluation and
+    # does not need mesh information
+
+    if evalf==True:
[…]
