### 2025-07-07 0c6ed17
Adding back nanoflann code temporarily
Why ?????

Because we did not code all the functions and so I wanted to have them back to make sure we don't lose the necessary code
