### 2025-07-07 f1daed6
Swarm -> evalf rework
Advection routines: evalf option is now just a flag passed in rather than a separate branch.
