+++ src/underworld3/swarm.py
@@ -808,7 +808,7 @@ class IndexSwarmVariable(SwarmVariable):
 ## be a PIC swarm
 
 
-class Swarm(Stateful, uw_object):
+class PICSwarm(Stateful, uw_object):
     """
     Particle swarm implementation with automatic mesh-particle interactions.
 
@@ -816,7 +816,7 @@ class Swarm(Stateful, uw_object):
     DMSWARM_PIC type. It provides automatic particle migration, mesh-particle connectivity,
     and streamlined particle operations for Lagrangian particle tracking and data storage.
 
-    Differences from UW_Swarm:
+    Differences from UW Swarm:
     - **Mesh Integration**: Built-in particle-in-cell (PIC) connectivity with automatic cell tracking
     - **Migration**: Uses the standard PETSc strategy for migration which depends on the DM type. This requires
       calculation of cell-relationships each time the coordinates are updated and particles that are not found will
@@ -1852,16 +1852,18 @@ class Swarm(Stateful, uw_object):
                 with self.access(self.particle_coordinates):
                     v_at_Vpts = np.zeros_like(self.particle_coordinates.data)
 
-                    if evalf:
-                        for d in range(self.dim):
-                            v_at_Vpts[:, d] = uw.function.evalf(
-                                V_fn_matrix[d], self.particle_coordinates.data
-                            ).reshape(-1)
-                    else:
-                        for d in range(self.dim):
-                            v_at_Vpts[:, d] = uw.function.evaluate(
-                                V_fn_matrix[d], self.particle_coordinates.data
-                            ).reshape(-1)
+                    # if evalf:
+                    #     for d in range(self.dim):
+                    #         v_at_Vpts[:, d] = uw.function.evalf(
+                    #             V_fn_matrix[d], self.particle_coordinates.data
+                    #         ).reshape(-1)
+                    # else:
+                    for d in range(self.dim):
+                        v_at_Vpts[:, d] = uw.function.evaluate(
+                            V_fn_matrix[d],
+                            self.particle_coordinates.data,
+                            evalf=evalf,
+                        ).reshape(-1)
 
                     mid_pt_coords = (
                         self.particle_coordinates.data[...]
@@ -1881,16 +1883,19 @@ class Swarm(Stateful, uw_object):
 
                     v_at_Vpts = np.zeros_like(self.data)
 
-                    if evalf:
-                        for d in range(self.dim):
-                            v_at_Vpts[:, d] = uw.function.evalf(
-                                V_fn_matrix[d], self.particle_coordinates.data
-                            ).reshape(-1)
-                    else:
-                        for d in range(self.dim):
-                            v_at_Vpts[:, d] = uw.function.evaluate(
-                                V_fn_matrix[d], self.particle_coordinates.data
-                            ).reshape(-1)
+                    # if evalf:
+                    #     for d in range(self.dim):
+                    #         v_at_Vpts[:, d] = uw.function.evalf(
+                    #             V_fn_matrix[d], self.particle_coordinates.data
+                    #         ).reshape(-1)
+                    # else:
+                    #
+                    for d in range(self.dim):
+                        v_at_Vpts[:, d] = uw.function.evaluate(
+                            V_fn_matrix[d],
+                            self.particle_coordinates.data,
+                            evalf=evalf,
+                        ).reshape(-1)
 
                     # if (uw.mpi.rank == 0):
                     #     print("Re-launch from X0", flush=True)
@@ -1911,16 +1916,18 @@ class Swarm(Stateful, uw_object):
[…]
