### 2025-07-08 6bc7b86
Added get_dof_partition() function for getting DOFs in each processor
