### 2025-07-08 5f33fb3
Adding rbf as alternative to evalf in function.evaluate
Also, fixes for example notebooks
