### 2025-07-09 4081c31
Added option of not saving DOF viewer outputs in a file.
