### 2025-07-24 f141acb
Update github action: build+test+deploy dev docker
