### 2025-07-24 2111529
Update setup.py
do not return from within a context manager
