### 2025-07-24 861bc10
Add conditional update to GH action
