+++ .github/workflows/docker-image.yml
@@ -1,19 +1,15 @@
 name: Image Build and Push
 
-# gh action to build and push container of uw3 to dockerhub
-# only runs on the 'development' branch for now. 
-# In future: run on 'main'/release, or any PR for testing.
-
 on:
   workflow_run: # dependency on build_uw3_and_test.yml
-    workflows: [test_uw3]
+    workflows: ["test_uw3"]
     types: [completed]
     branches: [development]
 
 jobs:
   push-to-dockerhub:
-    runs-on: ubuntu-latest
     if: ${{ github.event.workflow_run.conclusion == 'success' }}
+    runs-on: ubuntu-latest
 
     steps:
       - name: Checkout repository
@@ -36,5 +32,4 @@ jobs:
           file: ./Dockerfile
           platforms: linux/amd64
           # see https://github.com/docker/build-push-action/issues/276 for syntax help
-          tags: underworldcode/underworld3:${{ env.BRANCH }}
-          #-$(date +%s)
+          tags: underworldcode/underworld3:${{ env.BRANCH }} #-$(date +%s)
