### 2025-07-24 603bcf5
Update docker-image.yml
