### 2025-07-24 a43af52
Just make a dev docker with no dependency of other workflows
