### 2025-07-24 9a71dc6
Update syntax in Dockerfile
