### 2025-07-25 7482989
some small corrections to PR 2
