### 2025-07-28 #7831 @lmoresi
Starting our response to the outstanding review matters. I'll sign off and mention the review team when all the response comments are completed.

## Item: Updates to Quickstart Guide

Reviewer 2 asked us to include more detail on the parallel performance including a timing suite as part of the quick-start guide. We have added a section to explain parallel computing and a strong scaling graph (1-1000 CPUs). To be honest, parallel scaling is so dependent on architecture, HPC system policy, and problem that we aren't keen to write too much more for fear of being misleading.

The figure is also available online here: https://www.underworldcode.org/how-many-processors-should-we-use-to-solve-problem-x/?ref=underworld-geodynamics-community-newsletter and we expect to update these blog posts over time because people always ask for more examples.

Reviewer 2 asked us to include information on PETSc options. We have some basic examples in the quick-start guide and have added a link in the Stokes example to the PETSc tutorial on setting up solvers for Stokes. I am not keen to add too much here either because this is a genuine deep dive into the mathematics of iterative solvers and this guide is supposed to introduce the broad concepts within Underworld rather than give specific details. It is more appropriate for people to publish their scripts alongside their scientific results and include solver configurations there. If somebody does supply a PETSc solver configuration for their problem then it can be added quickly to the python / underworld solver configuration.

Reviewer 2 asked us to add the mathematical formulations, implementation strategies and timestepping details to the mathematical notebooks. What we have done is to include the commands that access the mathematical introspection in the  code to show the equations, time stepping details and references and we have improved the quality of the explanations built into the solvers (the documentation within the code is continually improving in response to questions / suggestions). The Navier Stokes example has nested objects to build the solver. Time-stepping is handled through `uw.ddt` objects for Eulerian, Semi-Lagrangian and fully Lagrangian implementations. This is the subject of a paper in preparation so we simply provide the code for the introspection.

Reviewer 2 commented that the adaptive meshing example is difficult to see. True (even if you zoom in) so we've rerun that example and have a single plume with a larger-scale gradient. There is not much need to plot the original mesh, so we leave that out as well.

NOTE - more responses to follow, no need to comment yet.
