### 2025-07-28 bd209c4
Changes to notebooks and documentation
Some rewriting of the internal solver documentation in response to JOSS reviews.

Updating notebooks to current codebase (e.g. evalf -> evaluate with rbf argument).

Improving adaptivity example
