### 2025-07-28 #7831 @lmoresi
## Item: Update to Paper

Reviewer 2 asked us to update the paper to include a description of the geodynamic context in the *statement of need* section.
