### 2025-07-28 #7831 @lmoresi
## Item: Fast-Forward codebase for JOSS branch (checklist)

There have been a number of breaking changes to the codebase associated with recent updates to PETSc but also important bug-fixes that we rely upon. It proved difficult for us to maintain the original installation instructions and API documentation for the JOSS branch in the face of these changes. We have therefore rebuilt the joss-submission branch against the most recent updates of the code. There are some interface changes in the quick-start Notebooks that the reviewers may notice.
