+++ src/underworld3/discretisation.py
@@ -1884,13 +1884,13 @@ class Mesh(Stateful, uw_object):
             for pt in range(0, face_num_points):
 
                 outside_control_point = (
-                    1e-8 * normal + 0.8 * points[pt] + 0.2 * face_centroid
+                    1e-8 * normal + 0.8 * point_coords[pt] + 0.2 * face_centroid
                 )
                 control_points_list.append(outside_control_point)
                 control_point_sign_list.append(-1)
 
                 inside_control_point = (
-                    -1e-8 * normal + 0.8 * points[pt] + 0.2 * face_centroid
+                    -1e-8 * normal + 0.8 * point_coords[pt] + 0.2 * face_centroid
                 )
                 control_points_list.append(inside_control_point)
                 control_point_sign_list.append(1)
@@ -2034,7 +2034,7 @@ class Mesh(Stateful, uw_object):
 
         # Part 2 - try to find the lost points by walking nearby cells
 
-        num_local_cells = self._centroid_index.data_pts.shape[0]
+        num_local_cells = self._centroids.shape[0]
         num_testable_neighbours = min(num_local_cells, 50)
 
         dist2, closest_centroids = self._centroid_index.query(
@@ -3337,7 +3337,7 @@ def checkpoint_xdmf(
 """
 
     ## The mesh Var attributes
-    
+
     def get_cell_field_size(h5_filename, mesh_var):
         with h5py.File(h5_filename, 'r') as f:
             size = f[f'cell_fields/{mesh_var.clean_name}_{mesh_var.clean_name}'].shape[0]
