### 2025-07-30 c63cf30
Fix array size mismatch when only one coord to evaluate is inside the mesh
