### 2025-07-31 8a9ec0a
adding some verbosity to the discretisation routines
