### 2025-07-31 dedaf0c
add launch script
