### 2025-08-01 16cddf5
kdtree swap out (temporary)
The pykdtree routines are currently hanging (on query) when using binder. Backing off to previous implementation (more bugs await, no doubt)
