### 2025-08-01 62d3bad
Update ckdtree.pyx
Make the cKDTree class switchable (i.e. same name as the pykdtree interface)
