### 2025-08-01 8076d7f
Some issues with inconsistency between old and new kdtree.
Fixing the old (C++ / hpp) one to match the new one and to use Julian's einstein summation speedups.
