+++ src/underworld3/ckdtree.pyx
@@ -7,7 +7,6 @@ import numpy as np
 
 from libcpp cimport bool
 
-
 cdef extern from "kdtree_interface.hpp" nogil:
     cdef cppclass KDTree_Interface:
         KDTree_Interface()
@@ -65,10 +64,20 @@ cdef class KDTree:
             raise RuntimeError(f"Provided points array dimensionality must be 2 or 3, not {points.shape[1]}.")
         self.points = points
         self.index = new KDTree_Interface(<const double *> &points[0][0], points.shape[0], points.shape[1])
+
         super().__init__()
 
     def __dealloc__(self):
         del self.index
+
+    @property
+    def n(self):
+        return self.points.shape[0]
+
+    @property
+    def ndim(self):
+        return self.points.shape[1]
+
 
     @timing.routine_timer_decorator
     def build_index(self):
@@ -187,18 +196,23 @@ cdef class KDTree:
         # return numpy data
         return n_indices, n_dist_sqr
 
+
     @timing.routine_timer_decorator
     def query(self,
-                  const double[: ,::1] coords not None:   numpy.ndarray,
-                  const int k                    :   numpy.int64,
-                  bool sqr_dists=True,
-
+             coords,
+             k=1,
+             sqr_dists=True,
     ):
         """
         Find the n points closest to the provided coordinates.
         """
 
-        i, d = self.find_closest_n_points(k, coords)
+        coords_contiguous = np.ascontiguousarray(coords)
+        i, d = self.find_closest_n_points(k, coords_contiguous)
+
+        # For consistency with pykdtree
+        if k==1:
+            i = i.reshape(-1)
 
         if sqr_dists:
             return numpy.sqrt(d), i
@@ -216,16 +230,15 @@ cdef class KDTree:
             coords,
             data,
             nnn = 4,
+            p=2,
             verbose = False,
         ):
 
         return self.rbf_interpolator_local_from_kdtree(
-            coords, data, nnn, verbose,
+            coords, data, nnn, p, verbose,
         )
 
-
-
-    def rbf_interpolator_local_from_kdtree(self,
+    def old_rbf_interpolator_local_from_kdtree(self,
             coords,
             data,
             nnn = 4,
@@ -286,7 +299,7 @@ cdef class KDTree:
 
         return Values
 
-    def rbf_interpolator_local_to_kdtree(self,
+    def old_rbf_interpolator_local_to_kdtree(self,
                     coords,
                     data,
                     nnn = 4,
@@ -352,3 +365,70 @@ cdef class KDTree:
         del Weights
 
         return Values
+
+
+    def rbf_interpolator_local_from_kdtree(self, coords, data, nnn, p, verbose):
+        """
+        Performs an inverse distance (squared) mapping of data to the target `coords`.
+        User can controls the algorithm by altering the number of neighbours used, `nnn` or the
+        power factor `p` of the mapping weighting.
+
+        Args:
+        coords  : ndarray,
+                The target spatial coordinates to evaluate the data from.
+                coords.shape[1] == self.ndim
+        data    : ndarray
+                The known data to map from. Must be full described over kd-tree.
+                i.e., data.shape[0] == self.n
+        nnn     : int,
+                The number of neighbour points to sample from, if `1` no distance averaging is done.
+        p       : int,
+                The power index to calculate weights, ie. pow(distance, -p)
+        verbose : bool,
+                Print when mapping occurs
+        """
+
+        if coords.shape[1] != self.ndim:
+            raise RuntimeError(
+                f"Interpolation coordinates dimensionality ({coords.shape[1]}) is different to kD-tree dimensionality ({self.ndim})."
+            )
[…]
