### 2025-08-02 4322355
Update src/underworld3/discretisation.py
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
