### 2025-08-04 6a98470
joss paper acknowledgement update
Add pawsey acknowledgement
update BK affiliation
