### 2025-08-05 2ce0e91
New Containerfile and container launch script.
This assumes containers will be run with `podman` and only in 'rootless
podman' mode.

This more is much better than 'rootful podman' or docker.
