### 2025-08-05 e146a3c
remove obsolete comment
