### 2025-08-05 774340f
Update build_uw3_and_test.yaml
