### 2025-08-05 9e4a325
Reactivate cache environment.yaml for GH workflow.
