### 2025-08-08 43c2020
Evaluation fixes
