+++ src/underworld3/function/_function.pyx
@@ -1,3 +1,4 @@
+from mpi4py.MPI import DATATYPE_NULL
 from libc.stdlib cimport malloc, free
 from typing import Optional, Tuple, Union
 
@@ -164,6 +165,84 @@ class UnderworldFunction(sympy.Function):
         return ourcls
 
 
+
+def global_evaluate(   expr,
+                np.ndarray coords=None,
+                coord_sys=None,
+                other_arguments=None,
+                simplify=True,
+                verbose=False,
+                evalf=False,
+                rbf=False,
+                data_layout=None,):
+    """
+    Evaluate a given expression at a list of coordinates - parallel-safe version
+
+    Note it is not efficient to call this function to evaluate an expression at
+    a single coordinate. Instead the user should provide a numpy array of all
+    coordinates requiring evaluation.
+
+    Parameters
+    ----------
+    expr: sympy.Basic
+        Sympy expression requiring evaluation.
+    coords: numpy.ndarray
+        Numpy array of coordinates to evaluate expression at.
+    coord_sys: mesh.N vector coordinate system
+
+    other_arguments: dict
+        Dictionary of other arguments necessary to evaluate function.
+        Not yet implemented.
+
+    """
+
+    mesh, varfns = uw.expressions.mesh_vars_in_expression(expr)
+
+    if mesh is None:
+        print(f"No mesh needed in evaluation - no need for parallel, global mesh variable shenanigans")
+        return
+
+    if uw.mpi.size == 1 or mesh is None:
+        return evaluate(
+            expr,coords,
+            coord_sys,
+            other_arguments,
+            simplify,
+            verbose,
+            evalf,
+            rbf,
+            data_layout,
+        )
+
+    # If in parallel, define a swarm, migrate, evaluate, migrate back
+    # (this is the routine used in advection - see ddt.c / SemiLagrangian)
+
+    # Set up a swarm, add a variable to represent the result of the computation and an 'original_rank' variable
+    # so that we can recover the information. We should add a local-index variable so we know how to reorder the
+    # values when the particles come back.
+
+    sample_swarm = uw.swarm.Swarm(mesh)
+
+    original_rank = uw.swarm.SwarmVariable(
+        "rank",
+        sample_swarm,
+        vtype=uw.VarType.SCALAR,
+        dtype=int,
+        _proxy=False,
+        varsymbol=r"\cal{R}",
+    )
+
+    # We need to know what variable type expr returns
+    # This should be wrapped up for everyone to use !
+
+    test_evaluation = evaluate(expr , np.atleast_2d(mesh._get_domain_centroids()[uw.mpi.rank]),
+                               reshape_1d=reshape_1d,
+)
+    eval_shape = np.atleast_2d(test_evaluation).shape
+
+    return
+
+
 def evaluate(   expr,
                 np.ndarray coords=None,
                 coord_sys=None,
@@ -171,7 +250,8 @@ def evaluate(   expr,
                 simplify=True,
                 verbose=False,
                 evalf=False,
-                rbf=False,):
+                rbf=False,
+                data_layout=None):
     """
     Evaluate a given expression at a list of coordinates.
 
@@ -190,85 +270,32 @@ def evaluate(   expr,
     other_arguments: dict
         Dictionary of other arguments necessary to evaluate function.
         Not yet implemented.
-
-
     """
 
-
-    # Extract all the mesh/swarm variables in the expression and check that they all live on the
-    # same mesh. If not, this evaluation is not valid.
-
-    varfns = set()
-    def unpack_var_fns(exp):
-
-        if isinstance(exp,uw.function._function.UnderworldAppliedFunctionDeriv):
[…]
