### 2025-08-08 f9ffdd4
Data structures for swarm and mesh variables
Wrapper function for the access manager to improve usability
