+++ src/underworld3/discretisation.py
@@ -896,7 +896,7 @@ class Mesh(Stateful, uw_object):
 
     def nuke_coords_and_rebuild(
         self,
-        verbose,
+        verbose=False,
     ):
         # This is a reversion to the old version (3.15 compatible which seems to work in 3.16 too)
         #
@@ -1867,12 +1867,7 @@ class Mesh(Stateful, uw_object):
                 - ((control_points_i - points) ** 2).sum(axis=1)
             ) > 0
 
-            # print(f"CPO {f}, {((control_points_o - points) ** 2).sum(axis=1)}")
-            # print(f"CPI {f}, {((control_points_i - points) ** 2).sum(axis=1)}")
             insiders[:, f] = inside[:]
-
-            # print(f"{f},  {inside}")
-            # print(f"{f},  {insiders}")
 
         return numpy.all(insiders, axis=1)
 
@@ -2083,8 +2078,6 @@ class Mesh(Stateful, uw_object):
                 )
         else:
             return np.zeros((0,))
-
-        print(f"Closest points {closest_points.shape}", flush=True)
 
         # We need to filter points that lie outside the mesh but
         # still are allocated a nearby element by this distance-only check.
@@ -2656,6 +2649,9 @@ class _MeshVariable(Stateful, uw_object):
         self.mesh.vars[self.clean_name] = self
         self._setup_ds()
 
+        # Setup public view of data
+        self._array = self._array_data_structure(self)
+
         return
 
     def _object_viewer(self):
@@ -2701,23 +2697,63 @@ class _MeshVariable(Stateful, uw_object):
 
         return newMeshVariable
 
-    def __getitem__(self, indices):
-        if not isinstance(indices, tuple):
-            if isinstance(indices, int) and self.shape[0] == 1:
-                i = 0
-                j = indices
-            else:
-                raise IndexError(
-                    "MeshVariable[i,j] access requires one or two indices "
-                )
+    # def __getitem__(self, indices):
+    #     if not isinstance(indices, tuple):
+    #         if isinstance(indices, int) and self.shape[0] == 1:
+    #             i = 0
+    #             j = indices
+    #         else:
+    #             raise IndexError(
+    #                 "MeshVariable[i,j] access requires one or two indices "
+    #             )
+    #     else:
+    #         i, j = indices
+
+    #     return self._data_container[i, j]
+
+    def pack(self, data_array):
+        """Convert an array in the correct shape for the underlying variable into something that can be loaded into
+        the flat storage structure used by PETSc in a numpy assigment (with index broadcasting etc)
+        """
+
+        import numpy as np
+
+        shape = self.shape
+        storage_size = self._data_layout(-1)
+        data_array_2d = np.atleast_2d(data_array)
+
+        with self.mesh.access(self):
+            for i in range(shape[0]):
+                for j in range(shape[1]):
+                    ij = self._data_layout(i, j)
+                    self.data[:, ij] = data_array_2d[:, i, j]
+
+        return
+
+    def unpack(self, squeeze=True):
+        """Return an array in the correct shape for the underlying variable from
+        the flat storage structure used by PETSc. By default, use numpy squeeze to remove additional
+        dimensions (keep those dimensions to leave all data as 3D array - scalars being shape (1,1), vectors
+        being (1,dim) and so on)
+        """
+
+        import numpy as np
+
+        shape = self.shape
+
+        with self.mesh.access(self):
+            points = self._data.shape[0]
+            data_array_3d = np.empty(shape=(points, *shape))
+
+            for i in range(shape[0]):
+                for j in range(shape[1]):
+                    ij = self._data_layout(i, j)
+                    data_array_3d[:, i, j] = self._data[:, ij]
[…]
