+++ src/underworld3/swarm.py
@@ -173,6 +173,7 @@ class SwarmVariable(Stateful, uw_object):
             )
 
         self._data = None
+        self._cached_data = None
         # add to swarms dict
 
         self.swarm._vars[self.clean_name] = self
@@ -204,51 +205,72 @@ class SwarmVariable(Stateful, uw_object):
 
         super().__init__()
 
+        self._array = self._array_data_structure(self)
+
         return
 
-    def __getitem__(self, indices):
-        if not isinstance(indices, tuple):
-            if isinstance(indices, int) and self.shape[0] == 1:
-                i = 0
-                j = indices
-            else:
-                raise IndexError(
-                    "SwarmVariable[i,j] access requires one or two indices "
-                )
-        else:
-            i, j = indices
+    # Deprecate this one
+    # def __getitem__(self, indices):
+    #     if not isinstance(indices, tuple):
+    #         if isinstance(indices, int) and self.shape[0] == 1:
+    #             i = 0
+    #             j = indices
+    #         else:
+    #             raise IndexError(
+    #                 "SwarmVariable[i,j] access requires one or two indices "
+    #             )
+    #     else:
+    #         i, j = indices
 
-        return self._data_container[i, j]
+    #     return self._data_container[i, j]
 
-    ## Should be a single master copy
+    ## Should be a single master copy (mesh variable / swarm variable)
     def _data_layout(self, i, j=None):
         # mapping
 
         if self.vtype == uw.VarType.SCALAR:
             return 0
         if self.vtype == uw.VarType.VECTOR:
-            if j is None:
-                return i
-            elif i == 0:
-                return j
+            if i < 0 or j < 0:
+                return self.swarm.dim
             else:
-                raise IndexError(
-                    f"Vectors have shape {self.mesh.dim} or {(1, self.mesh.dim)} "
-                )
+                if j is None:
+                    return i
+                elif i == 0:
+                    return j
+                else:
+                    raise IndexError(
+                        f"Vectors have shape {self.swarm.dim} or {(1, self.swarm.dim)} "
+                    )
         if self.vtype == uw.VarType.TENSOR:
-            if self.swarm.mesh.dim == 2:
-                return ((0, 1), (2, 3))[i][j]
+            if self.swarm.dim == 2:
+                if i < 0 or j < 0:
+                    return 4
+                else:
+                    return ((0, 1), (2, 3))[i][j]
             else:
-                return ((0, 1, 2), (3, 4, 5), (6, 7, 8))[i][j]
+                if i < 0 or j < 0:
+                    return 9
+                else:
+                    return ((0, 1, 2), (3, 4, 5), (6, 7, 8))[i][j]
 
         if self.vtype == uw.VarType.SYM_TENSOR:
-            if self.swarm.mesh.dim == 2:
-                return ((0, 2), (2, 1))[i][j]
+            if self.swarm.dim == 2:
+                if i < 0 or j < 0:
+                    return 3
+                else:
+                    return ((0, 2), (2, 1))[i][j]
             else:
-                return ((0, 3, 4), (3, 1, 5), (4, 5, 2))[i][j]
+                if i < 0 or j < 0:
+                    return 6
+                else:
+                    return ((0, 3, 4), (3, 1, 5), (4, 5, 2))[i][j]
 
         if self.vtype == uw.VarType.MATRIX:
-            return i + j * self.shape[0]
+            if i < 0 or j < 0:
+                return self.shape[0] * self.shape[1]
+            else:
+                return i + j * self.shape[0]
 
     def _create_proxy_variable(self):
         # release if defined
@@ -359,6 +381,70 @@ class SwarmVariable(Stateful, uw_object):
 
         return
 
+    # Need to be able to unpack as well
+    def pack(self, data_array):
[…]
