+++ environment.yaml
@@ -19,17 +19,23 @@ dependencies:
   - ipython
   - mpi4py
   - h5py=*=mpi*
-  - trame
-  - ipywidgets
   - gmsh
   - python-gmsh
   - python-xxhash
-  - jupyterlab
   - jupytext
-  - trame-vtk
-  - trame-vuetify
   - pygments
   - pint
+  - cmocean
+  - colorcet
+  - imageio
+  - imageio-ffmpeg
+  - ipywidgets<9.0.0
+  - jupyter-server-proxy
+  - jupyterlab<5.0.0
+  - trame>=2.5.2
+  - trame-vtk>=2.5.8
+  - trame-vuetify>=2.3.1
+  - trimesh
   - typeguard
   - pip
   - pip:
