### 2025-08-10 #7831 @diehlpk
@lmoresi please check the above invalid DOI.
