### 2025-08-10 #7831 @diehlpk
@lmoresi I read the paper and have the following comments

- Line 20: (?) seems to be a missing reference for sympy.
- Line 98: Not mandatory but here is a solution to make C++ look nicer in LaTeX (https://isocpp.org/wiki/faq/misc-environmental-issues#latex-macros)
- Line 129:  (https://doi.org/10.48569/18sb-8s4) -> Please make this a hyperlink, so readers can click on it
- Line 135: A full stop is missing at the end of the line.

Please address these comments and look at the DOI, so we can proceed with the acceptance of the paper.
