+++ tests/test_1120_SLVectorCartesian.py
@@ -14,12 +14,12 @@ nsteps = 20
 velocity = 1 / res  # /res
 
 # ### mesh coordinates
-xmin, xmax = 0., 1.
-ymin, ymax = 0., 1.
+xmin, xmax = 0.0, 1.0
+ymin, ymax = 0.0, 1.0
 
 sdev = 0.03
-x0 = 0.5*xmax
-y0 = 0.5*ymax
+x0 = 0.5 * xmax
+y0 = 0.5 * ymax
 
 # ### Set up the mesh
 ### Quads
@@ -38,6 +38,7 @@ unstructured_simplex_box_regular = uw.meshing.UnstructuredSimplexBox(
     cellSize=1 / res, regular=True, qdegree=3, refinement=0
 )
 
+
 @pytest.mark.parametrize(
     "mesh",
     [
@@ -50,26 +51,26 @@ def test_SLVec_boxmesh(mesh):
     print(f"Mesh - Coordinates: {mesh.CoordinateSystem.type}")
 
     # Create mesh vars
-    Vdeg        = 2
-    sl_order    = 2
+    Vdeg = 2
+    sl_order = 2
 
-    v           = uw.discretisation.MeshVariable("U", mesh, mesh.dim, degree=Vdeg)
-    vect_test   = uw.discretisation.MeshVariable("A", mesh, mesh.dim, degree=Vdeg)
+    v = uw.discretisation.MeshVariable("U", mesh, mesh.dim, degree=Vdeg)
+    vect_test = uw.discretisation.MeshVariable("A", mesh, mesh.dim, degree=Vdeg)
 
     # #### Create the SL object
     DuDt = uw.systems.ddt.SemiLagrangian(
-                                            mesh,
-                                            vect_test.sym,
-                                            v.sym,
-                                            vtype = uw.VarType.VECTOR,
-                                            degree = Vdeg,
-                                            continuous = vect_test.continuous,
-                                            varsymbol = vect_test.symbol,
-                                            verbose = False,
-                                            bcs = None,
-                                            order = sl_order,
-                                            smoothing = 0.0,
-                                        )
+        mesh,
+        vect_test.sym,
+        v.sym,
+        vtype=uw.VarType.VECTOR,
+        degree=Vdeg,
+        continuous=vect_test.continuous,
+        varsymbol=vect_test.symbol,
+        verbose=False,
+        bcs=None,
+        order=sl_order,
+        smoothing=0.0,
+    )
 
     # ### Set up:
     # - Velocity field
@@ -77,13 +78,17 @@ def test_SLVec_boxmesh(mesh):
     with mesh.access(v):
         v.data[:, 1] = velocity
 
-    x,y = mesh.X
+    x, y = mesh.X
 
     # vector components based on 2D Gaussian
     with mesh.access(vect_test):
-        gauss_2D             = sympy.exp(-((x - x0)**2 + (y - y0)**2)/(2*sdev**2))
-        vect_test.data[:, 0] = 2*velocity*uw.function.evaluate(gauss_2D, vect_test.coords)
-        vect_test.data[:, 1] = 2*velocity*uw.function.evaluate(gauss_2D, vect_test.coords)
+        gauss_2D = sympy.exp(-((x - x0) ** 2 + (y - y0) ** 2) / (2 * sdev**2))
+        vect_test.data[:, 0] = (
+            2 * velocity * uw.function.evaluate(gauss_2D, vect_test.coords).squeeze()
+        )
+        vect_test.data[:, 1] = (
[…]
