### 2025-08-12 39ef2ff
Update paper.md
In response to editor review
