### 2025-08-12 #7831 @lmoresi
@diehlpk

 - [x] I have found / fixed the doi for Zienkievicz's textbook https://doi.org/10.1016/C2009-0-24909-9.
 - [x] C++ ... yes, that works (even as a comment inside a reference - cool !!)
 - [x] I have fixed the refererence to sympy so this should now generate a valid link in the document
 - [x] I have fixed the link in the acknowledgements.
 - [x] Added the full stop.

Changes uploaded / committed to joss-submission branch.
