+++ .github/.devcontainer/uw_environment.yaml
@@ -1,1 +1,1 @@
-../../environment.yml
+../../environment.yaml
