+++ .github/workflows/envs/build_uw_docs.yaml
@@ -1,1 +1,1 @@
-../../../environment.yml
+../../../environment.yaml
