### 2025-08-12 17ffba3
Adding a README.md
Some useful links and overall desciption of the container technology
used.
