### 2025-08-12 802e48a
Adding a README.md
Some useful links and overall desciption of the container technology
used.
