### 2025-08-13 #7831 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
