### 2025-08-13 88947fe
Update paper.md
