### 2025-08-13 #7831 @lmoresi
That's fine. I've checked the paper.md in the branch and it is good to go.
